% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ab3}
\alias{ab3}
\title{Arab Baromteter Wave III Survey}
\format{
A data frame with 12 rows and 6 columns:
}
\usage{
ab3
}
\description{
A dataset containing estimates for twelve countries on a Likert-type rating
of confidence in the future economy. Respondents were asked,
"What do you think will be the economic situation in your country during the
next few years (3-5 years) compared to the current situation?"

Dataset is obtained from http://rnotr.com/likert/ggplot/barometer/likert-plots/.
}
\keyword{datasets}
