% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_bar_label_pos.R
\name{add_bar_label_pos}
\alias{add_bar_label_pos}
\title{Create a function that adds a column of label positions to a data frame.}
\usage{
add_bar_label_pos(dat)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
a data frame with one extra column indicating bar label positions.
}
\description{
\code{add_bar_label_pos()} is used to prepare data before making bar charts.
It takes a data frame as input and returns the same data frame with one extra
column of the mid points of the y variable (often Frequency or Percent that is
displayed on the y-axis in a bar chart) grouped by the x variable (a
categorical variable displayed on the x-axis in a bar chart).
}
\examples{
# make some fake data
year = c(rep(c("2006-07", "2007-08", "2008-09", "2009-10"), each = 4))
cat  = c(rep(c("A", "B", "C", "D"), times = 4))
freq = c(168, 259, 226, 340, 216, 431, 319, 368, 423, 645, 234, 685, 166, 467, 274, 251)
data = data.frame(year, cat, freq)
data

f = add_bar_label_pos(data)
data = f("year", "freq")
data
}
\seealso{
\code{\link{mk_barplot}}.
}

