% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data2.R
\name{agg_data2}
\alias{agg_data2}
\title{Aggregates data}
\usage{
agg_data2(data, cols = names(data), group_by = NULL,
  agg_fun = function(x) sum(x, na.rm = TRUE), group_by2 = NULL)
}
\description{
Aggregates data
}
\examples{
agg_data2(mtcars, "cyl")
agg_data2(mtcars, c("factor(cyl)", "cyl"))
agg_data2(mtcars, c(x = "factor(cyl)", y = "cyl"))
agg_data2(mtcars, c(x = "cyl", y = "cyl"), 1)
}
