% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_plot.R
\name{area_plot}
\alias{area_plot}
\title{area_plot}
\usage{
area_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  size = 20, palette = ez_col, ylabels = ez_labels,
  use_theme = theme_ez)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{group}{character}

\item{facet_x}{character}

\item{facet_y}{character}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}
}
\value{
A ggplot object.
}
\description{
area_plot
}
\examples{
area_plot(fruit, "OBS", "Units")
area_plot(fruit, "OBS", "Units", size = 14)
area_plot(fruit, "OBS", c("Weekly Units Sold" = "Units"))
area_plot(fruit, "OBS", "Units", "Product")
area_plot(fruit, "OBS", "Units", "Product", "Size")
area_plot(fruit, "OBS", "Units", "Product", "Size", "Store")
area_plot(fruit, "OBS", "Units", use_theme = ggplot2::theme_bw)
}
