% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attacks_all_countries}
\alias{attacks_all_countries}
\title{Internet attacks data, aggregated over all countries}
\format{
A data frame with 168 rows and 3 columns:
\describe{
     \item{wkday}{Week day: Monday, Tuesday, ..., Saturday, Sunday.}
     \item{hour}{00 - 24 hour of a day.}
     \item{n}{Number of attacks.}
}
}
\source{
\url{https://github.com/hrbrmstr/facetedcountryheatmaps/tree/master/data}
}
\usage{
attacks_all_countries
}
\description{
A dataset containing logs of internet attacks of all countries for each hour
of a day during a fixed period.
}
\keyword{datasets}
