% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attacks_by_country}
\alias{attacks_by_country}
\title{Internet attacks data by each country}
\format{
A data frame with 1680 rows and 5 columns:
\describe{
     \item{country}{Country name.}
     \item{wkday}{Week day: Monday, Tuesday, ..., Saturday, Sunday.}
     \item{hour}{00 - 24 hour of a day.}
     \item{country_code}{2-letter country code.}
     \item{n}{Number of attacks.}
}
}
\source{
\url{https://github.com/hrbrmstr/facetedcountryheatmaps/tree/master/data}
}
\usage{
attacks_by_country
}
\description{
A dataset containing logs of internet attacks by country for each hour
of a day during a fixed period.
}
\keyword{datasets}
