% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births}
\alias{births}
\title{A dataset of information at birth for 44 babies}
\format{
A data frame with 44 rows and 4 variables, sorted in ascending order
of birth time. (The time window is 24 hours, i.e., 44 babies were born in 24
hours.)
\describe{
     \item{sex}{Gender of the baby.}
     \item{weight_g}{Weight (in grams) at birth.}
     \item{minutes}{Time of birth converted to minutes since midnight.}
     \item{diffs}{Difference (in minutes) between consecutive birth times.}
}
}
\source{
Dunn, "A Simple Dataset for Demonstrating Common Distributions," Journal of Statistics Education v.7, n.3 (1999)
}
\usage{
births
}
\description{
On December 18, 1997, 44 babies were born in a hospital in Brisbane, Australia.
This dataset contains their birth times and other information at birth.
}
\keyword{datasets}
