% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btbo_by_year}
\alias{btbo_by_year}
\title{Aggregated Annual Budget and Boxoffice}
\format{A data frame with 202 rows and 4 variables:
\describe{
     \item{year}{Year of release.}
     \item{type}{"Budget" or "Boxoffice".}
     \item{tot}{Total Amount in US Dollars.}
     \item{avg}{Average Amount in US Dollars.}
}}
\usage{
btbo_by_year
}
\description{
A dataset containing the ggregated annual Budget and Boxoffice of
5944 films beween 1913 and 2014 from IMDB.com.
}
\keyword{datasets}

