% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binom_ci.R
\name{calc_binom_ci}
\alias{calc_binom_ci}
\title{Calculate a binomial proportion confidence interval from a sample}
\usage{
calc_binom_ci(x, method = "prop.test", conf.level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of [0, 1] values representing the sample}

\item{method}{Which method to use to construct the interval. Must be one of
"exact", "ac", "asymptotic", "wilson", "prop.test", "bayes", "logit",
"cloglog", "probit". See `binom` package for details. Default is
"prop.test".}

\item{conf.level}{The specified confidence level. Defatuls is 0.95.}

\item{na.rm}{If `TRUE`, NA values will be removed from the sample. If
`FALSE`, NA values will be set to 0. Default is `TRUE`.}
}
\description{
Uses any of various techniques for calculating a binomial
  proportion confidence interval from a sample using the `binom` package
}
