% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer Survival Rates}
\format{A data frame with 96 rows and 3 variables:
\describe{
     \item{group}{Different cancer types}
     \item{year}{x years since cancer onset}
     \item{value}{survival rates, proportion of patients survived}
}}
\usage{
cancer
}
\description{
A dataset containing the survival rates of 24 different cancers.
}
\keyword{datasets}

