% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_colors.R
\docType{data}
\name{cb_colors}
\alias{cb_colors}
\title{Colorblind-friendly color palette}
\format{
An object of class \code{character} of length 9.
}
\source{
\url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}
}
\usage{
cb_colors
}
\description{
A palette of 9 colorblind-friendly Hex values developed by Masataka Okabe
and Kei Ito.
}
\examples{
cb_colors['reddish purple']; names(cb_colors)
}
\keyword{datasets}
