% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{combine_plots}
\alias{combine_plots}
\title{Combines several ggplot objects (plots) into one figure.}
\usage{
combine_plots(
  ...,
  align = "v",
  ncol = 2,
  labels = "auto",
  label_size = 10,
  title = "Grand Title",
  title_size = 12
)
}
\arguments{
\item{...}{Several ggplot objects. Pass in no more than 6 to avoid crowdness.}

\item{align}{Specifies whether the plots should be horizontally ("h") or
vertically ("v") aligned. Options are "none", "h", "v" (default), and "hv"
(align in both directions).}

\item{ncol}{Number of columns in the plot grid. Default = 2.}

\item{labels}{List of labels to be added to the plots. You can also set
\code{labels="auto"} (default) to auto-generate lower-case labels (a. b. c. d. ...),
\code{labels="AUTO"} for upper-case labels (A. B. C. D. ...), or
\code{labels='autonum'} for integer labels (1. 2. 3. 4. ...).}
}
\value{
A ggplot object that shows multiple graphs in one figure.
}
\description{
Sometimes we want to put several (ggplot) plots together in one figure for
publication. \code{combine_plots} does exactly that, allowing additional
parameters to control the alignment and index labels of the individual plots,
and etc.
}
\examples{
inst/examples/ex-combine_plots.R
}
