% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{common}
\alias{common}
\title{The default colors in ggplot2 can be difficult to distinguish from one another
because they have equal luminance. They are also not friendly for colorblind
viewers. A good general-purpose solution is to just use the
colorblind-friendly palette below taken directly from
http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}
\format{\preformatted{ chr [1:7] "#E69F00" "#56B4E9" "#009E73" "#F0E442" ...
}}
\usage{
common
}
\description{
cbPalette is the palette with gray, and cbbPalette is with black.
}
\keyword{datasets}

