% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days_in_office}
\alias{days_in_office}
\title{U.S. Presidents Days in Office}
\format{
A data frame with 44 rows and 2 columns:
\describe{
     \item{president}{Name of the president.}
     \item{days}{Number of days in office.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_presidents_of_the_United_States_by_time_in_office}
}
\usage{
days_in_office
}
\description{
A dataset containing the number of days in office for all U.S. presidents
until 2020.
}
\keyword{datasets}
