% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(
  data,
  x,
  group = NULL,
  facet_x = NULL,
  facet_y = NULL,
  adjust = 1,
  alpha = 0.5,
  facet_scales = "fixed",
  ncol = NULL,
  env = parent.frame()
)
}
\description{
creates a densityi plot
}
\examples{
density_plot(mtcars, "wt", "cyl")
density_plot(subset(tsibbledata::olympic_running, Length == "100m" & Year >= 1980),
             "Time", "Year - Year \%\% 10", "Sex", facet_scales = "free", ncol = 1, adjust = 2)
}
