% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params_paretodist.R
\name{est_params_paretodist}
\alias{est_params_paretodist}
\title{Estimate the shape and location parameters of the pareto distribution
from which the sample data were taken.}
\usage{
est_params_paretodist(df, varname, digits = 3)
}
\arguments{
\item{df}{A data frame.}

\item{varname.}{String, name of a continuous variable with an empirical
distribution approximately pareto.}

\item{digits.}{Integer, the number of digits after decimal point to keep for
the estimated parameter values.}
}
\value{
Estimates of the shape and location parameters of a theoretical
pareto distribution.
}
\description{
Given a data frame and the name of a continuous variable from the data frame,
suppose the CCDF of the variable on a log10-log10 scale is linear, then its
distribution can be modeled by the pareto distribution, of which the shape
and location parameters can be estimated by this function.
}
\examples{
library(ezplot)
est_params_paretodist(pops, 'pop', digits = 2)
}
