% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params_weibull.R
\name{est_params_weibull}
\alias{est_params_weibull}
\title{Estimate the parameters of the weibull distribution from which the
sample data were taken.}
\usage{
est_params_weibull(df, varname, digits = 3)
}
\arguments{
\item{df}{A data frame.}

\item{varname.}{String, name of a continuous variable with an empirical
distribution approximately weibull.}

\item{digits.}{Integer, the number of digits after decimal point to keep for
the estimated parameter values.}
}
\value{
Estimates of the shape (k) and location (lambda) parameters of the
weibull distribution.
}
\description{
Given a data frame and the name of a continuous variable from the data frame,
suppose the -logCCDF of the variable on a log-log scale is linear, then its
distribution can be modeled by the weibull distribution, of which the
parameters can be estimated by this function.
}
\examples{
library(ezplot)
est_params_weibull(births, 'diffs')
est_params_weibull(births, 'diffs', digits = 4)
}
