% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ests_CIs}
\alias{ests_CIs}
\title{Fake data for testing mk_forestplot()}
\format{
A data frame with 26 rows and 5 columns:
}
\usage{
ests_CIs
}
\description{
A dataset containing fake point estimates and their confidence intervals of
different groups. See inst/examples/ex-mk_forestplot.R
}
\keyword{datasets}
