% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\name{ex_data}
\alias{ex_data}
\title{example data}
\usage{
ex_data(start_date = as.Date("2012-01-01"),
  end_date = as.Date("2016-12-31"), char = c("A", "B"),
  fct = factor(c("X", "Y", "Other"), c("X", "Y", "Other")), num = c(10, 20))
}
\description{
example data
}
\examples{
ex_data()
}
