% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_app.R
\name{ez_app}
\alias{ez_app}
\title{ez_app}
\usage{
ez_app(data = NULL)
}
\arguments{
\item{data}{A data frame}
}
\description{
ez_app
}
\examples{
\dontrun{
library(tsibble)
library(tsibbledata)
ez_app(ansett)
}
}
