% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_col.R
\name{ez_col}
\alias{ez_col}
\title{Color palette interpolation}
\usage{
ez_col(n = 50, palette = NULL)
}
\arguments{
\item{n}{number of colours}

\item{palette}{palette to interpolate from}
}
\value{
rgb
}
\description{
Color palette interpolation
}
\examples{
ez_col(15)
ez_col(2, c("blue", "red"))
ez_col(3, c("blue", "red"))
}
