% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_data.R
\name{ez_data}
\alias{ez_data}
\title{ez_data}
\usage{
ez_data(start_date = as.Date("2012-01-01"),
  end_date = as.Date("2016-12-31"), char = c("A", "B"),
  fct = factor(c("X", "Y", "Other"), c("X", "Y", "Other")), num = c(10, 20),
  seed = 1)
}
\arguments{
\item{start_date}{A start date value.}

\item{end_date}{An end date value.}

\item{char}{Vector of character values.}

\item{fct}{Vector of factor values.}

\item{num}{Vector of numeric values.}

\item{seed}{A random seed.}
}
\value{
A data.frame.
}
\description{
Creates example data using \code{expand.grid()} on \code{char},
  \code{fct} and \code{num} and then adding random \code{units} and
  \code{value} columns.
}
\examples{
df = ez_data()
summary(df)
}
