% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_data.R
\name{ez_data2}
\alias{ez_data2}
\title{ez_data2}
\usage{
ez_data2(start_date = as.Date("2012-01-01"),
  end_date = as.Date("2016-12-31"), xsec = c("apples", "oranges"),
  seed = 9)
}
\arguments{
\item{start_date}{A start date value.}

\item{end_date}{An end date value.}

\item{xsec}{(optional) vector of cross sections.}

\item{seed}{A random seed.}
}
\description{
ez_data2
}
