% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_col.R
\name{ez_jet}
\alias{ez_jet}
\title{ez_jet}
\usage{
ez_jet(
  n = 100,
  palette = c("dodgerblue4", "steelblue2", "olivedrab3", "darkgoldenrod1", "brown")
)
}
\arguments{
\item{n}{Number of colours to return.}

\item{palette}{Vector of colours.}
}
\description{
ez_jet
}
