% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_log_breaks.R
\name{ez_log_breaks}
\alias{ez_log_breaks}
\title{ez_log_breaks}
\usage{
ez_log_breaks(limits)
}
\arguments{
\item{limits}{Limits of axis}
}
\description{
Creates break vector for a log axis
}
\examples{
ez_log_breaks(limits = c(0.9, 11))
}
