% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ezline}
\alias{ezline}
\title{Shortcuts}
\usage{
ezline(data, aggr = NULL, x, y = NULL, group = NULL,
  position = "identity", facet = NULL, facet_x = NULL, facet_y = NULL,
  palette = NULL, ci = NULL, error_type = "errorbar", error_args = NULL,
  text = FALSE, text_round = function(x) signif(x, 3), text_args = NULL,
  title = NULL, subtitle = NULL, caption = NULL, ...)
}
\description{
List of convenience functions that call ezplot with common defaults
pre-supplied (e.g. plot_type)
}
