% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezplot.R
\name{ezplot}
\alias{ezplot}
\title{Interactive tool that creates well-formated charts and graphs with one simple function.}
\usage{
ezplot(data, x = NA, y = NA, type = NA)
}
\arguments{
\item{data}{A table of data}

\item{x}{The column/row name of the data input that will be plotted on the x-axis}

\item{y}{The column/row name of the data input that will be plotted on the y-axis}

\item{type}{The type of graph/chart that the user wants returned}
}
\value{
a graph of the data input in the specified type and appropriate summary statistics
}
\description{
Interactive tool that creates well-formated charts and graphs with one simple function.
}
\examples{
ezplot(iris)
ezplot(iris, "Sepal.Length", "Sepal.Width", "scattergram")
ezplot(iris, "Sepal.Length", "Sepal.Width", "line")
ezplot(mpg, "hwy", "None", "bar")
ezplot(iris, "Sepal.Length", "Sepal.Width", "smooth")
ezplot(mpg, "class", "hwy", "boxplot")
ezplot(mpg, "hwy", "None", "frequency")
ezplot(mpg, "hwy", "None", "histogram")
ezplot(economics, "date", "unemploy", "area")
}

