% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruit}
\alias{fruit}
\title{Fruit sales data}
\format{A data frame with 1872 rows and 6 variables:
\describe{
  \item{OBS}{Weekly dates}
  \item{Product}{item sold}
  \item{Size}{size}
  \item{Store}{store}
  \item{Units}{number of units sold}
  \item{Value}{revenue}
}}
\usage{
fruit
}
\description{
data.frame for time series plots and aggregation
}
\keyword{datasets}
