% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{gantt_plot}
\alias{gantt_plot}
\title{Create Gantt Chart}
\usage{
gantt_plot(data, task, start, end, category, sprint.start = NULL,
  sprint.end = NULL, progress = NULL)
}
\arguments{
\item{data}{Data frame}

\item{task}{Column name giving task descriptions}

\item{start}{Column name giving column containing start dates (in `POSIXct` format)}

\item{end}{Column name giving column containing start dates (in `POSIXct` format)}

\item{category}{(optional) Column name defining categories to be used to
split the gantt chart}

\item{sprint.start}{(optional) Start date for the sprint (`POSIXct`)}

\item{sprint.end}{(optional) End date for the sprint (`POSIXct`)}

\item{progress}{(optional) Column name giving progress as a percentage.
Will only be used if sprint start and end are provided.}
}
\description{
Function to create a ggplot2 gantt chart
}
