% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bar_labels_freq.R
\name{get_bar_labels_freq}
\alias{get_bar_labels_freq}
\title{Helper function for \code{mk_barplot_freq}. Not for external use.}
\usage{
get_bar_labels_freq(df, var, fillby, show_pct)
}
\arguments{
\item{df}{data frame of the original data where bar charts are made of.}

\item{var}{string, name of the categorical variable to be counted.}

\item{fillby}{string, name of a second categorical variable used for
sub-dividing and coloring the bars.}

\item{show_pct}{logical, if TRUE, calculate bars middle position using pct.
Otherwise, calculate bars middle position using count. Default is FALSE.}
}
\value{
a data frame containing bar labels and their positions
}
\description{
\code{get_bar_labels_freq} takes in the original data frame, and
returns a data frame of labels and their postions for \code{mk_barplot_freq}.
}
\seealso{
\code{\link{mk_barplot_freq}}.
}
