% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bar_labels_resp.R
\name{get_bar_labels_resp}
\alias{get_bar_labels_resp}
\title{Helper function for \code{mk_barplot_resp}. Not for external use.}
\usage{
get_bar_labels_resp(df, gp, resp, fillby)
}
\arguments{
\item{df}{data frame of the original data where bar charts are made of.}

\item{gp}{string, name of a categorical variable to group the response var.}

\item{resp}{string, name of a continuous variable as the response var.}

\item{fillby}{string, name of a second categorical variable used for
sub-dividing and coloring the bars.}
}
\value{
a data frame containing bar labels and their positions
}
\description{
\code{get_bar_labels_resp} takes in the original data frame, and
returns a data frame of labels and their postions for \code{mk_barplot_resp}.
}
\seealso{
\code{\link{mk_barplot_resp}}.
}
