% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bar_labels_yvar.R
\name{get_bar_labels_yvar}
\alias{get_bar_labels_yvar}
\title{Helper function for \code{barplot_yvar}. Not for external use.}
\usage{
get_bar_labels_yvar(df, xvar, yvar, fillby)
}
\arguments{
\item{df}{data frame of the original data where bar charts are made of.}

\item{xvar}{string, name of x categorical variable.}

\item{yvar}{string, name of y continuous variable.}

\item{fillby}{string, name of a second categorical variable used for
sub-dividing the (aggregated) y values and coloring the bars.}
}
\value{
a data frame containing bar labels and their positions
}
\description{
\code{get_bar_labels_yvar} takes in the original data frame, the
x, y and fillby variables. It returns a data frame of labels and their
postions for \code{barplot_yvar}.
}
\seealso{
\code{\link{mk_barplot_yvar}}.
}
