% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summ_stats.R
\name{get_summ_stats}
\alias{get_summ_stats}
\title{Helper function for \code{mk_histogram}. Not for external use.}
\usage{
get_summ_stats(df, xvar, facet_by)
}
\arguments{
\item{df}{original data frame.}

\item{xvar}{string, name of a continuous variable for x-axis.}

\item{facet_by}{string, name of a categorical variable for grouping x and
creating a facet for each group.}
}
\value{
a data frame containing summary stats for each group.
}
\description{
\code{get_summ_stats} takes in the original data frame, the
x and group variables. It returns a data frame of the summary stats (median,
mean and etc) of x in each group. Used to add lines at these summary stats
for histograms.
}
\seealso{
\code{\link{mk_histogram}}.
}
