% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{histogram_plot}
\usage{
histogram_plot(
  data,
  x,
  y = "count",
  group = NULL,
  facet_x = NULL,
  facet_y = NULL,
  position = "stack",
  bins = 30,
  alpha = 0.5,
  facet_scales = "fixed",
  ncol = NULL,
  env = parent.frame()
)
}
\description{
creates a densityi plot
}
\examples{
histogram_plot(airquality, "Wind", group = "Month")
histogram_plot(airquality, "Wind", "density", facet_x = "Month")
}
