% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_plot.R
\name{ks_plot}
\alias{ks_plot}
\title{ks_plot}
\usage{
ks_plot(
  data,
  fitted,
  actual,
  palette = ez_col,
  size_line = 1,
  size = 11,
  env = parent.frame()
)
}
\arguments{
\item{data}{A data.frame.}

\item{fitted}{Vector of fitted values}

\item{actual}{Vector of actual values}

\item{palette}{Colour function.}

\item{size_line}{width of line for \code{geom_line()}. Default is 1.}

\item{size}{theme size for \code{use_theme()}. Default is 14.}

\item{env}{environment for evaluating expressions.}
}
\description{
ks plot
}
\examples{
ks_plot(mtcars, "-disp", "am")
x = c(rnorm(100), rnorm(100) + 2)
label = c(rep('low', 100), rep('high', 100))
ks_plot(data.frame(x, label), "x", "label")
ks_plot(data.frame(x, label = factor(label, c('low', 'high'))), "x", "label")
}
