% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{line_plot}
\usage{
line_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  linesize = 1, size = 20, palette = ez_col, ylabels = ez_labels,
  use_theme = theme_ez, facet_scales = "fixed")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{group}{character}

\item{facet_x}{character}

\item{facet_y}{character}

\item{linesize}{width of line for \code{geom_line()}. Default is 1.}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}
}
\value{
A ggplot object.
}
\description{
Creates line plots.
}
\examples{
library(ukbabynames)
library(dplyr)
df = ukbabynames \%>\%
  filter(!grepl("^-", name)) \%>\%
  mutate(first_letter = toupper(substring(name, 1, 1)),
         vowel_start = first_letter \%in\% c("A", "E", "I", "O", "U"),
         name_length = ifelse(nchar(name) > 5, "Long", "Short"))
line_plot(df, "year", "n")
line_plot(df, "year", c("Number of Babies" = "n"))
line_plot(df, "year", c("Number of Babies" = "n"), "sex")
line_plot(df, "year", "n", "sex", "first_letter")
line_plot(df, "year", "n", "sex", "first_letter", facet_scales = "free_y")
line_plot(df, "year", "n", "sex", "name_length", "vowel_start")
line_plot(df, "year", "n", use_theme = ggplot2::theme_bw)
}
