% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{lineplot}
\alias{lineplot}
\title{lineplot}
\usage{
lineplot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL, ...)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{Bare expression.}

\item{y}{Bare expression.}

\item{group}{Bare expression.}

\item{facet_x}{Bare expression.}

\item{facet_y}{Bare expression.}

\item{...}{arguments to pass to \code{line_plot}}
}
\value{
ggplot
}
\description{
see \code{line_plot}.
}
\examples{
lineplot(fruit, OBS, Units)
}
