% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plot.R
\name{mean_plot}
\alias{mean_plot}
\title{mean_plot}
\usage{
mean_plot(data, x, y, size = 20, labels = ez_labels)
}
\description{
mean_plot
}
\examples{
library(dplyr)
tibble::rownames_to_column(mtcars, "car") \%>\%
  mean_plot("cyl", "hp > 110",
            labels = function(x) ez_labels(100 * x, append = "\%"))
}
