% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_fixer}
\alias{missing_fixer}
\title{Change missing value code to NA.}
\usage{
missing_fixer(na_value)
}
\arguments{
\item{na_value}{A number or a string.}
}
\value{
\code{function(x)}
\itemize{
     \item x     :  a numeric or character vector.
}
}
\description{
\code{missing_fixer} takes a number or a string that represents missing value
and returns a function that can be used to replace the number or string 
occurred in a vector with NA.
}
\examples{
fix_missing_99 = missing_fixer(-99)
fix_missing_999 = missing_fixer(-999)
fix_missing_char = missing_fixer("missing")

fix_missing_99(c(-99, -999))
fix_missing_999(c(-99, -999))
fix_missing_char(c("missing", "male", "female"))
}

