% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_barplot.R
\name{mk_barplot}
\alias{mk_barplot}
\title{Create a function that draws ggplot2 barplots}
\usage{
mk_barplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby, xorder, barpos,
               xlab="", ylab="", main="", legend=T,
               barlab, barlab_use_pct, decimals, barlab_at_top, barlab_size,
               dodged_lab_w, dodged_lab_h)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item fillby   :  string, the variable used for coloring the bars.
     \item xorder   :  string, "alphanumeric", "ascend" or "descend". It specifies how the categories are ordered on the x-axis. Default is "alphanumeric".
     \item barpos   :  string, "stack" or "dodge" bars. Default is "stack".
     \item xlab     :  string, the x-axis label.
     \item ylab     :  string, the y-axis label.
     \item main     :  string, the title of the plot.
     \item legend   :  logical, whether to show the legend. Default is TRUE.
     \item barlab   :  string, the name of the variable that contains the bar labels. Default is NULL.
     \item barlab_use_pct: logical, whether to use pct format for the bar labels. Default is FALSE.
     \item decimals      : number of decimal places for percent labels, default is 2.
     \item barlab_at_top : logical, whether to place the labels at the top or middle of the bars or. Default is FALSE, meaning at the middle.
     \item barlab_size   : numeric, the size of the bar label text. Default is 3.
     \item dodged_lab_w  : numeric, the width between the dodged bar labels. Default is 1.
     \item dodged_lab_h  : numeric, the height between the dodged bar labels. Default is 1. Often not used.
}
}
\description{
\code{mk_barplot} takes a data frame as input and returns a function that
can be used to make barplots using the variables in the data frame.
}
\examples{
# make some fake data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
barplt("student", "grade", "student")
barplt("student", "grade", "student", xorder="ascend", legend=F)

# make some fake data
df2 = read.table(header=TRUE, text='
group level val
A      small 1.8
A      medium 2.2
A      large 1.5
B      small 2.0
B      medium 2.6
B      large 1.0
C      small 2.5
C      medium 1.3
C      large 2.9')

# draw a barplot
barplt = mk_barplot(df2)
barplt("group", "val", "level")

# calculate the percentage of the levels within each group
library(tidyr)
library(dplyr)
pct = df2 \%>\% spread(level, val)
temp = pct[, 2:4]
pct = cbind(group=pct[, 1], temp / apply(temp, 1, sum))
pct = pct \%>\% gather(level, pct, -group)
pct$level = factor(pct$level, levels=c("small", "medium", "large"))

# plot a stacked barplot to display the percentages
barplt = mk_barplot(pct)
p = barplt("group", "pct", "level")
scale_axis(p, "y", scale="pct")
scale_axis(p, "y", scale="pct", pct_jump=0.1)
}
\seealso{
\code{\link{scale_axis}} for adding different scales to the axes.
         \code{\link{add_bar_label_pos}} for adding label positions to the input data frame.
}

