% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barplot_freq.R
\name{mk_barplot_freq}
\alias{mk_barplot_freq}
\title{Create a function for making publishable ggplot2 barplots, showing
frequency counts (or percentages) of a categorical variable or 2 categorical
variables.}
\usage{
mk_barplot_freq(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, fillby = "1", xorder = "alphanumeric", show_pct = FALSE,
               label_decimals = 1, label_size = 3, legend_title = fillby,
               font_size = 14)}
\itemize{
     \item xvar     :  string, name of a categorical variable for x-axis.
     \item fillby   :  string, name of a different categorical variable for
                       subdividing and coloring the bars.
                       Default is "1", meaning no such variable is supplied.
     \item xorder   :  string, "alphanumeric", "ascend" or "descend". It
                       specifies how categories are ordered on the x-axis.
                       Default is "alphanumeric".
     \item show_pct :  logical, if TRUE, show percent on y-axis;
                       otherwise, show count. Default is FALSE.
     \item label_decimals: integer, the number of decimal points shown
                       on the bar labels. Default = 1.
     \item label_size: integer, size of bar label text. Default is 3.
                       Hide bar labels when its value is 0.
     \item legend_title: string, legend title. Default is the name of the
                         fillby variable.
     \item font_size:  overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_barplot_freq} takes a data frame as input and returns a function for
making bar charts with any categorical variable from the data frame on the
x-axis and frequency counts (or percents) of its levels on the y-axis.
The output function can also produce dodged (for counts) or stacked
(for percents) bar charts when supplied a second categorical variable, a
fillby variable. The resulting bar chart will have bars ordered by the
alphanumerical order of the x levels and labeled with both counts and
percents by default. Plus, it'll have a clean theme and clean fonts.
It'll also use color-blind friendly palettes.
}
\examples{
inst/examples/ex-mk_barplot_freq.R
}
\seealso{
\code{\link{scale_axis}} for adding different scales to the y-axis.
}
