% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barplot_tufte.R
\name{mk_barplot_tufte}
\alias{mk_barplot_tufte}
\title{Create a function that draws tufte bar charts.}
\usage{
mk_barplot_tufte(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, gpvar, xlab="", ylab="", main="",
               pt_size=8, font_size=2.5, ...)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot. 
     \item gpvar    :  string, the grouping variable.
     \item pt_size  :  numeric, size of points. Default = 8
     \item font_size:  numeric, size of text. Default = 2.5
     \item ...      :  parameters (base_size or base_family) to pass into 
                       theme_slopegraph()
}
}
\description{
Inspired and modified after Lukasz Piwek's article at 
http://motioninsocial.com/tufte/#minimal-barchart
}

