% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barploth_freq.R
\name{mk_barploth_freq}
\alias{mk_barploth_freq}
\title{Create a function for making publishable ggplot2 horizontal barplots,
showing frequency counts (or percentages) of a categorical variable or 2
categorical variables.}
\usage{
mk_barploth_freq(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(yvar, fillby = "1", yorder = "alphanumeric", show_pct = FALSE,
               label_size = 3, font_size = 14)}
\itemize{
     \item yvar     :  string, name of a categorical variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       subdividing and coloring the bars.
                       Default is "1", meaning no such variable is supplied.
     \item yorder   :  string, "alphanumeric", "ascend" or "descend". It
                       specifies how categories are ordered on the y-axis.
                       Default is "alphanumeric".
     \item show_pct :  logical, if TRUE, show percent on y-axis;
                       otherwise, show count. Default is FALSE.
     \item label_size: integer, size of bar label text. Default is 3.
                       Hide bar labels when its value is 0.
     \item font_size:  overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_barploth_freq} takes a data frame as input and returns a function
for making horizontal bar charts with any categorical variable from the data
frame on the y-axis and frequency counts (or percents) of its levels on the
x-axis. The output function can also produce dodged (for counts) or stacked
(for percents) bar charts when supplied a second categorical variable, a
fillby variable. The resulting bar chart will have bars ordered by the
alphanumerical order of the y levels by default, and bars
labeled with both counts and percents. Plus,
it'll have a clean theme and clear fonts.
It'll also use color-blind friendly palettes.
}
\examples{
library(ezplot)
f = mk_barploth_freq(films)

f("mpaa")
f("mpaa", yorder = "descend")
f("year_cat", font_size = 10)
f("year_cat", fillby = "made_money")

# use label_size = 0 to remove the bar labels
f("year_cat", fillby = "made_money", yorder = "descend", label_size = 0)
f("mpaa", fillby = "made_money")
f("made_money", fillby = "year_cat")
f("made_money", fillby = "year_cat", yorder = "descend", label_size = 0)
f("year_cat", fillby = "mpaa")

f("mpaa", show_pct = T)
f("mpaa", show_pct = T, yorder = "descend")
f("mpaa", fillby = "made_money", show_pct = T)
f("mpaa", fillby = "made_money", show_pct = T, yorder = "descend")
f("year_cat", fillby = "made_money", show_pct = T)
f("year_cat", fillby = "mpaa", show_pct = T)

f = mk_barploth_freq(ggplot2::diamonds)
f("cut")
f("cut", show_pct = T, font_size = 9)
f("cut", show_pct = T, fillby = "clarity") +
    ggplot2::theme(legend.position = "bottom") +
    ggplot2::guides(fill = ggplot2::guide_legend(nrow = 1))

# use label_size = 0 to remove the bar labels
p = f("cut", fillby = "clarity", font_size = 11, label_size = 0)
add_labs(p, title = "Fuel efficiency generally decreases with engine size",
         subtitle = "Two seaters (sports cars) are an exception ...",
         caption = "Data from fueleconomy.gov")
}
\seealso{
\code{\link{scale_axis}} for adding different scales to the y-axis.
}
