% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_cdfplot.R
\name{mk_cdfplot}
\alias{mk_cdfplot}
\title{Create a function for making publishable ggplot2 CDF plots.}
\usage{
mk_cdfplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, colorby = "1", complement = FALSE, linew = 0.7,
               add_vline_median = FALSE, show_label_median = TRUE,
               legend_title = colorby, legend_pos = "right",
               label_size = 3, font_size = 14)}
\itemize{
     \item xvar. String, name of a continuous variable for x-axis.
     \item colorby. String, name of a categorical variable for grouping x.
     For each group, a CDF curve will be drawn with a different color.
     Default = "1", indicating no such variable is supplied.
     \item complement. Logical (default = FALSE), if TRUE, produce a function
     for plotting the complement of CDF.
     \item linew. Number, width of the line. Default = 0.7.
     \item add_vline_median. Logical (default = FALSE), if TRUE, add vertical
     dashed line segments at the median values going up and touching the
     CDF curves. To make it easy to look at, also add horizontal dashed line
     at y=0.5.
     \item show_label_median. Logical (default = TRUE), if TRUE, show median
     values along the vertical median lines. Otherwise, don't show.
     \item legend_title. String, legend title. Default is the name of the
     colorby variable.
     \item legend_pos. String, legend position. Default = "right".
     \item label_size. Integer, size of bar label text. Default = 3.
     \item font_size. Overall font size. Default = 14. The font size of the
     axes and legend text is a fraction of this value.
     \item .... Optional parameters for `stat_ecfd()`. For example, instead
     of using the default `geom = 'step'`, you can set `geom = 'point'`,
     which will draw the ecdf in dots instead of stepped line segments. For
     another example, instead of using the default `pad = TRUE`, you can set
     `pad = FALSE` to stop the ecdf curve at its boundaries given by the data
     instead of extending it to +Inf and -Inf horizontally.
}
}
\description{
\code{mk_cdfplot} takes a data frame as input and returns a function for
plotting empirical cumulative distribution function (ECDF) of any variable
(continuous or categorical) from the data frame.
}
\examples{
inst/examples/ex-mk_cdfplot.R
}
