% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mk_distplot.R
\name{mk_distplot}
\alias{mk_distplot}
\title{Create a function that draws ggplot2 histograms or density plots.}
\usage{
mk_distplot(df)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
\code{function(xvar, fillby="", xlab="", type="histogram", binw=NULL,
               main="", xlog=F, xlog10=F, xpct=F, xpct_jump=0.2)}
\itemize{
     \item xvar     :  string, the x variable.
     \item fillby   :  string, the grouping variable. Default is "".
     \item type     :  string, "density" or "histogram". Default is "histogram".
     \item binw     :  number, bin width. Default is NULL.
     \item xlab     :  string, the x-axis label.
     \item main     :  string, the title of the plot.
     \item xlog     :  logical, indicating whether to use log scale on x-axis. Default is FALSE.
     \item xlog10   :  logical, indicating whether to use log10 scale on x-axis. Default is FALSE.
     \item xpct     :  logical, indicating whether to use percent format on x-axis. Default is FALSE.
     \item xpct_jump:  numeric, between 0 and 1. Default is 0.2
}
}
\description{
\code{mk_distplot()} takes a data frame as input and returns a function that
can be used to make histograms or density plots on variables in the data frame.
}
\examples{
f = mk_distplot(iris)
f("Sepal.Length")
f("Sepal.Length", binw=0.3)
f("Sepal.Length", type="density")

f("Sepal.Length", fillby="Species")
f("Sepal.Length", fillby="Species", binw=0.3)
f("Sepal.Length", fillby="Species", type="density")
}

