% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_facet_lineplot.R
\name{mk_facet_lineplot}
\alias{mk_facet_lineplot}
\title{Create a function for making publishable ggplot2 faceted plot.}
\usage{
mk_facet_lineplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, xvar_top, yvar_rt, gpby = "1", ylab_rt = NULL,
               legend_title = gpby, palette = "D", linew = 0.7,
               font_size = 14)}
\itemize{
     \item xvar    : string, name of a continuous var for the bottom x-axis.
     \item yvar    : string, name of a continuous var for the left y-axis.
     \item xvar_top: string, name of a categorical var for the top x-axis.
     \item yvar_rt : string, name of a continuous var for the right y-axis.
     \item gpby    : string, name of a continuous or categorical variable for
                     coloring the curves. Default = "1", meaning no such
                     variable is supplied.
     \item ylab_rt : string, 2nd y-axis label. If NULL (default), use the
                     value of yvar_rt.
     \item legend_title: string, legend title. Default is the name of the
                         gpby variable.
     \item palette : string, the colormap option to use. Possible values are
                     "A", "B", "C", "D" (default) and "E".
     \item linew   : number, width of the line. Default = 0.7.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_facet_lineplot} takes a data frame as input and returns a function for
making faceted plots that can display relationship of upto 5 variables at
one time.
}
\examples{
inst/examples/ex-mk_facet_lineplot.R
}
