% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_heatmap.R
\name{mk_heatmap}
\alias{mk_heatmap}
\title{Create a function that draws ggplot2 heat maps.}
\usage{
mk_heatmap(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby, xlab="", ylab="", main="", base_size = 12, 
               use_theme_gray=TRUE, legend=TRUE)}
\itemize{
     \item xvar     :  string, x variable.
     \item yvar     :  string, y variable.
     \item fillby   :  string, variable used for coloring the bars.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot. 
     \item base_size:  numeric, size of tick text for both axes. Default = 12.
     \item use_theme_gray:  logical, whether to use theme_gray or theme_minimal. Default = TRUE.
     \item legend   :  logical, whether to show the legend. Default = TRUE.
}
}
\description{
\code{mk_heatmap} takes a data frame as input and returns a function that 
can be used to make heat maps using variables in the data frame.
}

