% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_intervalplot.R
\name{mk_intervalplot}
\alias{mk_intervalplot}
\title{Create a function that draws ggplot2 interval plot.}
\usage{
mk_intervalplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby="", ymin_var, ymax_var, xlab="", ylab="",
               main="", size=1)}
\itemize{
     \item xvar     :  string, x variable.
     \item yvar     :  string, y variable.
     \item fillby   :  string, variable used for coloring the points and lines.
     \item ymin_var :  string, variable name for ymin.
     \item ymax_var :  string, variable name for ymax.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot.
     \item size     :  number, size of the points and lines. Default = 1.
     \item legend   :  logical, whether to show the legend. Default is TRUE.
}
}
\description{
\code{mk_intervalplot} takes a data frame as input and returns a function that
can be used to make interval plots using variables in the data frame. An
interval plot is a chart that represent an interval as a vertical line, with
a point in the middle.
}
\examples{
library(ezplot)
plt = mk_intervalplot(films)
}

