% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_likertplot.R
\name{mk_likertplot}
\alias{mk_likertplot}
\title{Create a function for making publishable ggplot2 likert plot (a.k.a.,
horizontal diverging bar plot), showing values of a continuous variable by a
2 categorical variables.}
\usage{
mk_likertplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby, fillby_lvls, yorder = "alphanumeric",
               x_as_pct = FALSE, legend_title = fillby, font_size = 14)}
\itemize{
     \item xvar     :  string, name of a continuous variable for x-axis.
     \item yvar     :  string, name of a categorical variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       sub-dividing and coloring the bars.
     \item fillby_lvls: character vector, levels of fillby variable that the
                        fillby variable should be ordered accordingly.
     \item yorder   :  string, "alphanumeric", "ascend" or "descend". It
                       specifies how categories are ordered on the y-axis.
                       Default = "alphanumeric".
     \item x_as_pct :  logical, if TRUE, format x-axis as %;
                       otherwise, format it as comma. Default is FALSE.
     \item legend_title: string, legend title. Default is the name of the
                         fillby variable.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_likertplot} takes a data frame as input and returns a function for
making likert plots (a.k.a., horizontal diverging bar plots) with any
categorical variable from the data frame on the y-axis and the value of
any continuous variable on the x-axis, with bars colored by a 2nd
categorical fillby variable. The resulting plot can have bars ordered from
top to bottom in alphanumerical, ascending or descending order.
}
\examples{
inst/examples/ex-mk_likertplot.R
}
\seealso{
\code{\link{scale_axis}} for adding different scales to the x-axis.
}
