% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_slopegraph.r
\name{mk_slopegraph}
\alias{mk_slopegraph}
\title{Create a function that draws ggplot2 tufte slopegraphs.}
\usage{
mk_slopegraph(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, gpvar, xlab="", ylab="", main="",
               pt_size=8, font_size=2.5, ...)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot.
     \item gpvar    :  string, the grouping variable.
     \item pt_size  :  numeric, size of points. Default = 8
     \item font_size:  numeric, size of text. Default = 2.5
     \item ...      :  parameters (base_size or base_family) to pass into
                       theme_slopegraph()
}
}
\description{
Create a function that draws ggplot2 tufte slopegraphs.
}

