% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_plot.R
\name{na_plot}
\alias{na_plot}
\title{na_plot}
\usage{
na_plot(data, palette = ez_col)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A ggplot object.
}
\description{
Visual representation of the NAs in a data.frame
}
\examples{
na_plot(airquality)
}
