% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nameifnot}
\alias{nameifnot}
\title{nameifnot}
\usage{
nameifnot(x, make.names = FALSE)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A named vector.
}
\description{
Names unnamed elements of a character vector.
}
