% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_null}
\alias{no_null}
\title{no_null}
\usage{
no_null(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
y
}
\description{
Converts "NULL" character to NULL.
}
\examples{
no_null(NULL)
no_null("NULL")
no_null("NOPE")
}
