% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palette.R
\name{palette}
\alias{palette}
\title{Color palette that I use a lot in my work.}
\usage{
palette(name)
}
\arguments{
\item{name}{"cb_gray", "cb_black", "blue", "red", "purple", or "green".}
}
\value{
a character vector of color palettes
}
\description{
\code{palette()} takes in a color name and returns the corresponding palette.
}
\examples{
palette("cb_gray")
palette("blue")
}

