% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_plot.R
\name{perf}
\alias{perf}
\title{perf}
\usage{
perf(actual, fitted, y_value, x_value)
}
\arguments{
\item{actual}{Vector with two levels}

\item{fitted}{Vector with values between 0 and 1}
}
\description{
Precision recall calculation
}
\examples{
ezplot:::perf(sample(c(TRUE, FALSE), 1, replace = TRUE), runif(1), "lift", "rpp")
ezplot:::perf(sample(c(TRUE, FALSE), 10, replace = TRUE), runif(10), "lift", "rpp")
ezplot:::perf(sample(c(TRUE, FALSE), 5, replace = TRUE), runif(5), "prec", "rec")
ezplot:::perf(sample(c(TRUE, FALSE), 5, replace = TRUE), runif(5), "tpr", "fpr")
}
