% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_df}
\alias{perf_df}
\title{perf_df}
\usage{
perf_df(fitted, actual)
}
\arguments{
\item{fitted}{A character value. Evaluates to a numeric column.}

\item{actual}{A character value. Evaluates to a logical or binary column.}
}
\description{
perf_df
}
\examples{
perf_df(mtcars$mpg, mtcars$am)
perf_df(mtcars$wt, mtcars$am==0)
}
