% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_plot.R
\name{pie_plot}
\alias{pie_plot}
\title{pie_plot}
\usage{
pie_plot(data, x, y, facet_x = NULL, facet_y = NULL, ylabels = ez_labels,
  size = 20, label_cutoff = 0.04, round = Inf, signif = 3,
  palette = ez_col, label_x = 1.3)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{Value.}

\item{y}{Value.}

\item{facet_x}{Value.}

\item{facet_y}{Value.}

\item{ylabels}{Label formatting function.}

\item{size}{Theme size for \code{use_theme()}. Default is 20.}

\item{label_cutoff}{Label cutoff value.}

\item{round}{Option for rounding label.}

\item{signif}{Option for retaining significant figures in label.}

\item{palette}{Colour function.}

\item{label_x}{Position of label from centre of pie.}
}
\value{
ggplot object
}
\description{
pie_plot
}
\examples{
pie_plot(fruit, "Product", "Units")
pie_plot(fruit, "Product", "Units", "Store")
pie_plot(fruit, "Product", "Units", "Store", "Size")
}
