% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Waterfall plot}
\usage{
plot_waterfall(data, aggr = NULL, x, y, group = NULL, group_sign = FALSE,
  direction = "up", touch = NULL, palette = NULL)
}
\arguments{
\item{data}{(`data.frame` | `tibble`) Data to be used in plot}

\item{aggr}{(`function`) (Optional) function created using}

\item{x}{Variable to use on X-axis. Must be of type 'character' or 'factor'.}

\item{y}{Variable to use on Y-axis. Must be of type 'numeric' or 'integer'.}

\item{group}{(Optional) Variable to group by}

\item{group_sign}{(`logical`) (Optional) Group by sign (positive or negative)
for each segment? A new column will be created consisting of `pos`, `neg`,
`touch` for each segment (factor levels will be set in order listed).
If `TRUE`, will override `group`. Default is `FALSE`.}

\item{direction}{= Should the waterfall build 'up' or 'down'? Default is 'up'.}

\item{touch}{(Optional) Vector of x-variables which will touch the X-axis. Default is
the first and last variable in the sequence.}

\item{palette}{(Optional) Color palette to use for fill}
}
\description{
Produce a waterfall plot using the geom_rect() function. A waterfall plot
shows a sequence of numbers as an increase or decrease from the cumulative
sum of the previous numbers. It is useful in breaking out a change in values
into components and comparing the relative size and direction of each
component. For an example, see:
\url{https://learnr.wordpress.com/2010/05/10/ggplot2-waterfall-charts/}
}
