% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plt_dist.R
\name{plt_dist}
\alias{plt_dist}
\title{Create a figure of 4 plots to show the distribution of a numeric variable.}
\usage{
plt_dist(dat, varname)
}
\arguments{
\item{dat}{A data frame or a matrix with colnames.}

\item{varname}{A string specifying the name of the variable of interest.}
}
\description{
\code{plt_dist} takes in a data frame and a variable name, and plots its
histogram, boxplot, density curve and normal qqplot using the base plot system.
}
\examples{
plt_dist(cars, "speed")
}

