% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_n_ssize_gender}
\alias{power_n_ssize_gender}
\title{Fake data for testing mk_facet_lineplot()}
\format{
A data frame with 448 rows and 5 columns:
}
\usage{
power_n_ssize_gender
}
\description{
A dataset containing fake powers and sample sizes when testing main effect
between male and female. See inst/examples/ex-mk_facet_lineplot.R
}
\keyword{datasets}
