% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ppareto}
\alias{ppareto}
\title{CDF of the Pareto Distribution.}
\usage{
ppareto(q, shape = 1, location = 0)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shape}{vector of shapes. Default = 1.}

\item{location}{vector of locations. Default = 0.}
}
\value{
A vector of probabilities.
}
\description{
Cumulative Distribution Function for the pareto distribution with shape
`shape` and location `location`.
}
\examples{
ppareto(seq(1, 5, 1), shape = 3, location = 1)
}
