% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_dumbbell.R
\name{prep_data_dumbbell}
\alias{prep_data_dumbbell}
\title{Helper function for \code{mk_dumbbell}. Not for external use.}
\usage{
prep_data_dumbbell(df, xvar_left, xvar_right, yvar, yorder)
}
\arguments{
\item{df}{Data frame of the original data.}

\item{xvar_left}{String, name of a numeric variable, left end-point on x.}

\item{xvar_right}{String, name of a numeric variable, right end-point on x.}

\item{yvar}{String, name of y categorical (character or factor) variable.}

\item{yorder}{String, possible values are "alphanumeric", "descend" or
"ascend". If "alphanumeric", order y levels in alphanumerical order
along y-axis. If "ascend"/"descend", order y levels in ascending/descending
order of the difference between xvar_right and xvar_left.}
}
\value{
The original data frame with y ordered according to instruction and a
new column called diff appended.
}
\description{
\code{prep_data_dumbbell} takes in the original data frame, the
left and right end point x variables, the y variable, and the
instruction on how to order the y levels. It returns a data frame with y
ordered according to the supplied instruction, and a new column called
diff (difference between the right and left end point xvars) appended.
}
\seealso{
\code{\link{mk_dumbbell}}.
}
