% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-data-likert.R
\name{prep_data_likert}
\alias{prep_data_likert}
\title{Helper function for \code{mk_likertplot}. Not for external use.}
\usage{
prep_data_likert(df, xvar, yvar, fillby, fillby_lvls, yorder)
}
\arguments{
\item{df}{data frame of the original data.}

\item{xvar}{string, name of x numeric variable.}

\item{yvar}{string, name of y categorical (character or factor) variable.}

\item{fillby}{string, name of a second categorical variable for sub-dividing
and coloring the bars.}

\item{fillby_lvls}{character vector, levels of the fillby variable that the
fillby variable should be ordered accordingly.}

\item{yorder}{string, possible values are "alphanumeric", "descend" or
"ascend". If "alphanumeric", order y levels in alphanumerical order
along y-axis. If "ascend"/"descend", order y levels in ascending/descending
order of the sum of absolute x-values along y-axis.}
}
\value{
a list of 2 data frames (df_neg and df_pos), breaks and labels for
the continuous axis, and palette for coloring the bars (pal).
}
\description{
\code{prep_data_likert} takes in the original data frame, the
x, y and fillby variable, the supplied levels of the fillby variable, and
instruction on how to order the y levels. It returns a list of 2 data frames
(df_neg and df_pos), along with some plot elements for making a likert chart,
a.k.a., horizontal diverging bar chart.
}
\seealso{
\code{\link{mk_likertplot}}.
}
