% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-data-lollipop.R
\name{prep_data_lollipop}
\alias{prep_data_lollipop}
\title{Helper function for \code{mk_lollipop}. Not for external use.}
\usage{
prep_data_lollipop(df, xvar, yvar, yorder)
}
\arguments{
\item{df}{Data frame of the original data.}

\item{xvar}{String, name of x numeric variable.}

\item{yvar}{String, name of y categorical (character or factor) variable.}

\item{yorder}{String, possible values are "alphanumeric", "descend" or
"ascend". If "alphanumeric", order y levels in alphanumerical order
along y-axis. If "ascend"/"descend", order y levels in ascending/descending
order of the sum of absolute x-values along y-axis.}
}
\value{
A data frame for making and labeling a horizontal lollipop chart.
}
\description{
\code{prep_data_lollipop} takes in the original data frame, the
x, y variables, and instruction on how to order the y levels. It returns a
data frame for making and labeling a horizontal lollipop chart.
}
\seealso{
\code{\link{mk_lollipop}}.
}
