% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc}
\alias{roc}
\title{roc}
\usage{
roc(fitted, actual)
}
\arguments{
\item{fitted}{Vector with values between 0 and 1}

\item{actual}{Vector with two levels}
}
\description{
Calculates ROC and AUC
}
\examples{
ezplot:::roc(runif(1), sample(c(TRUE, FALSE), 1, replace = TRUE))
ezplot:::roc(runif(3), sample(c(TRUE, FALSE), 3, replace = TRUE))
}
