% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{roc_plot}
\usage{
roc_plot(actual, fitted, size = 20)
}
\description{
roc_plot
}
\examples{
actual = sample(c(0, 1), 100, replace = TRUE)
fitted = runif(100)
roc_plot(actual, fitted)

}
