% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_axis_text.R
\name{rotate_axis_text}
\alias{rotate_axis_text}
\title{Rotate axis tick text and adjust their distances to the axis.}
\usage{
rotate_axis_text(p, text_angle_x = 0, text_angle_y = 0,
  hjust_x = NULL, hjust_y = NULL, vjust_x = NULL, vjust_y = NULL)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{text_angle_x}{Angle of tick text on x-axis. Default = 0.}

\item{text_angle_y}{Angle of tick text on y-axis. Default = 0.}

\item{hjust_x}{Vertical distance adjustment of tick text to x-axis. Default = NULL.}

\item{hjust_y}{Horizontal distance adjustment of tick text to y-axis. Default = NULL.}

\item{vjust_x}{Horizontal distance adjustment of tick text to x-axis. Default = NULL.}

\item{vjust_y}{Vertical distance adjustment of tick text to y-axis. Default = NULL.}
}
\value{
A ggplot2 object.
}
\description{
Rotate axis tick text and adjust their distances to the axis.
}
\examples{
inst/examples/ex-rotate_axis_text.R
}
