% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_png.R
\name{save_png}
\alias{save_png}
\title{save_png}
\usage{
save_png(g, file, width, height, res, ..., vp = NULL)
}
\arguments{
\item{g}{A ggplot or ezplot object.}

\item{file}{A png file path.}

\item{width}{Width of output image.}

\item{height}{Height or output image.}

\item{res}{Resolution of output image.}

\item{...}{Further arguments to pass to \code{png()}.}

\item{vp}{A viewport object created with grid::viewport.}
}
\description{
Saves ggplot or ezplot objects to png.
}
