% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_axis.R
\name{scale_axis}
\alias{scale_axis}
\title{Use a different scale on either axis of a ggplot2 plot.}
\usage{
scale_axis(p, axis = "y", scale = "default", nticks = 10, digits)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{axis}{A string of value "x" or "y". Default = "y".}

\item{scale}{A string of value "default", "comma", "dollar", "pct", "log",
"log1p", "log10", "log2", "sqrt", or "exp". It specifies which scale to use.}

\item{nticks}{Number of ticks on axis. Default = 10. The actual number of
ticks shown will be 1 or 2 less or more that the number user supplied depends
on the actual limit of the data. This is caused by rounding in computation.}

\item{digits}{Number of digits after the decimal point when using the 'pct'
scale. Default uses best guess. It's only effective when `scale = 'pct'`.}
}
\value{
A ggplot2 object with the new scale applied to the input axis.
}
\description{
\code{scale_axis} takes a ggplot2 object as input and applies to its x-axis
or y-axis one of the following scales:
\itemize{
    \item comma . Show numbers in 000,000 format.
    \item dollar. Show $ in front of numbers.
    \item pct   . Use percent format.
    \item log   . Use log transformation.
    \item log1p . Use log(x+1) transformation.
    \item log10 . Use log10 transformation.
    \item log2  . Use log2 transformation.
    \item sqrt  . Use squared root transformation.
    \item exp   . Use exponential transformation.
}
If a ggplot object has too few breaks on an axis (to see the max value), and
if you don't need to apply any scales to the axis (due to the values are
already in a nice range), you can apply scale='default' to that axis, and
this will add nticks (a number supplied by you) from the min to the max value
on that axis.
}
\examples{
inst/examples/ex-scale_axis.R
}
