% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter plot}
\usage{
scatter_plot(data, x, y, group = NULL, size = 14, point_size = 2.5,
  env = parent.frame())
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character value. Evaluates to a column.}

\item{y}{A named character value. Evaluates to a column.}

\item{group}{A character value. Evaluates to a column.}

\item{size}{theme size for \code{use_theme()}. Default is 14.}

\item{point_size}{Numeric. Default is 2.}
}
\description{
create a scatter plot
}
\examples{
scatter_plot(mtcars, "wt", "hp")
scatter_plot(mtcars, "wt", "hp", "factor(cyl)")
scatter_plot(mtcars, "factor(cyl)", "hp")
}
