% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slopegraph_data_prep.R
\name{slopegraph_data_prep}
\alias{slopegraph_data_prep}
\title{Build a slopegraph data set. This function is inspired and modified after
James Keirstead's original script at https://github.com/jkeirstead/r-slopegraph}
\usage{
slopegraph_data_prep(df, xvar, yvar, gpvar, min_space = 0.05, ...)
}
\arguments{
\item{df}{A data frame.}

\item{xvar}{String, name of the x-axis column.}

\item{yvar}{String, name of the y-axis column.}

\item{gpvar}{String, name of the group column.}

\item{min_space}{Number, fraction of total data range to use as a minimum gap
                 between the lines. Default = 0.05.}
}
\value{
A data frame with an additional ypos column.
}
\description{
Modify a data frame so that it can be used to plot tufte slopegraph. A tufte
slopegraph has the following characteristics: values in the first x-column
are sorted based on their numeric value. Subsequent group lines are then
shifted to ensure that the lines for two adjacent groups never cross.
Vertical positions in subsequent columns are only meaningful relative to the
first entry in that group.
}
\details{
It uses the function \code{\link{tufte_sort}} internally, which does
the heavy lifting of the above mentioned tufte slopegraph format.
}
\seealso{
\code{\link{tufte_sort}}
}

