% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally1way}
\alias{tally1way}
\title{Calculates the freq and pct of each levels in a categorical variable 
       that exists in a data frame}
\usage{
tally1way(dat)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar)}
\itemize{
     \item xvar     :  string, name of a categorical variable
}
}
\description{
Before you plot a bar chart, you first need to count the frequencies of 
each level in the categorical variable. \code{tally1way} makes that easy. 
It takes a data frame as input and returns a function, which takes the 
name of the categorical variable of your interest that exists in the data  
frame, and returns the frequencies and percentages of the categories.
}
\examples{
data(films)
f = tally1way(films)
f("year_cat")
f("make_money")
}
\seealso{
\code{\link{mk_barplot}}.
}

