% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally2way}
\alias{tally2way}
\title{Calculates the freq and pct of each combination of the levels in 
       two categorical variables that exists in a data frame}
\usage{
tally2way(dat)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar)}
\itemize{
     \item xvar     :  string, name of a categorical variable
     \item yvar     :  string, name of a categorical variable
}
}
\description{
\code{tally2way} makes it really easy to calculate the frequencies and 
percentages of the combinations of the levels in two categorical variables.  
It takes a data frame as input and returns a function, which takes the 
names of any two categorical variables of your interest that exist in the data  
frame, and returns the frequencies and percentages of each of the combinations 
of the categories.
}
\examples{
data(films)
f = tally2way(films)
f("year_cat", "make_money")
}
\seealso{
\code{\link{mk_barplot}}.
}
