% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_expdist.R
\name{test_expdist}
\alias{test_expdist}
\title{Create a function for drawing publishable ggplot2 CDF and CCDF (on
log-y scale) plots side by side on one canvas. The resulting figure is used
to test if the data distribution is exponential.}
\usage{
test_expdist(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(varname, linew = 0.7, xlab = varname, title_left, title_right,
               subtitle_left, subtitle_right, caption_left, caption_right,
               xscale_left, ...)}
\itemize{
     \item varname. String, name of a continuous variable. Its empirical CDF
     will be plotted along side its complement CDF.
     \item linew. Number, width of the line. Default = 0.7.
     \item xlab. String, x label of the left and the right figures.
     Default is varname.
     \item title_left. String, title of the left figure.
     \item title_right. String, title of the right figure.
     \item subtitle_left. String, subtitle of the left figure.
     \item subtitle_right. String, subtitle of the right figure.
     \item caption_left. String, caption of the left figure.
     \item caption_right. String, caption of the right figure.
     \item xscale_left. String, scale of x-axis of the left figure. Possible
     values are described at \code{\link{scale_axis}}. There's no parameter
     to change the x-scale of the right figure because we need to keep the
     x-axis of CCDF at the original/raw scale in order to detect if the
     distribution is exponential or not.
     \item .... Other parameters for making a CDF plot. A common one, for
     example, is `add_vline_median = TRUE`, which will add a vertical line at
     the median. Another common one is `show_label_median = FALSE`, which
     will suppress the display of median value along the median vline. See
     \code{\link{mk_cdfplot}} for a full list of parameters.
}
}
\description{
\code{test_expdist} takes a data frame as input and returns a function for
making CDF and CCDF (on log-y scale) plots of any continuous variable
from the data frame. CCDF standands for Complement CDF. If CCDF on log-y
scale is linear, the observed data are exponentially distributed.
}
\examples{
inst/examples/ex-test_expdist.R
}
