% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Create a function for drawing publishable ggplot2 CDF plot and normal
probability plot (also called Q-Q normal plot) side by side on one canvas.
The resulting figure is often used to test normality of the sample distribution.}
\usage{
test_normality(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(varname, detrend = TRUE, title_left, title_right, ...)}
\itemize{
     \item varname. String, name of a continuous variable. Its empirical CDF
     will be plotted along side its normal probability plot.
     \item detrend. Logical (default = TRUE), If TRUE, the normal probability
     plot will be detrended according to the reference Q-Q line. As a result,
     if the data points are scattered around the horizontal line y = 0, then
     we can conclude the data are approximately normal. The detrend
     procedure was described by Thode (2002), and it reduces visual bias
     caused by orthogonal distances from Q-Q points to the reference line.
     \item title_left. String, title of the left figure.
     \item title_right. String, title of the right figure.
     \item .... Other parameters for making a CDF plot. A common one, for
     example, is `add_vline_median = TRUE`, which will add a vertical line at
     the median. Another common one is `show_label_median = FALSE`, which
     will suppress the display of median value along the median vline. See
     \code{\link{mk_cdfplot}} for a full list of parameters.
}
}
\description{
\code{test_normality} takes a data frame as input and returns a function for
drawing CDF plot and normal probability plot of any continuous variable
from the data frame.
}
\examples{
inst/examples/ex-test_normality.R
}
