% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_col.R
\name{text_contrast}
\alias{text_contrast}
\title{text_contrast}
\usage{
text_contrast(x)
}
\arguments{
\item{x}{Vector of colours.}
}
\value{
Vector indicating whether black or white should be used for text
  overlayed on x.
}
\description{
text_contrast
}
\examples{
text_contrast("#000000")
text_contrast("black")
}
