% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cowplot.R
\name{theme_cowplot}
\alias{theme_cowplot}
\title{Modified cowplot theme}
\usage{
theme_cowplot(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)
}
\arguments{
\item{font_size}{Overall font size.}

\item{font_family}{Font family for plot title, axis titles and labels,
legend texts, etc.}

\item{line_size}{Line size for axis lines.}

\item{rel_small}{Relative size of small text (e.g., axis tick labels)}

\item{rel_tiny}{Relative size of tiny text (e.g., caption)}

\item{rel_large}{Relative size of large text (e.g., title)}
}
\description{
Made right margin big enough to show max x-axis value for horizontal plot.
Made no backgroud when using facets.
See \url{https://github.com/wilkelab/cowplot/blob/master/R/themes.R} for
the original cowplot theme.
}
