% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ez.R
\name{theme_ez}
\alias{theme_ez}
\title{Default theme}
\usage{
theme_ez(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base fond family}
}
\value{
theme
}
\description{
Default theme
}
\examples{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(cyl, mpg)) + theme_ez()
}
