% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_no_xyaxes.R
\name{theme_no_xyaxes}
\alias{theme_no_xyaxes}
\title{A theme for ggplot2 heatmap}
\usage{
theme_no_xyaxes(font_size = 14, font_family = "", line_size = 0.5,
  rel_small = 12/14, rel_tiny = 11/14, rel_large = 16/14)
}
\arguments{
\item{font_size}{Overall font size.}

\item{font_family}{Font family for plot title, axis titles and labels,
legend texts, etc.}

\item{line_size}{Line size for axis lines.}

\item{rel_small}{Relative size of small text (e.g., axis tick labels)}

\item{rel_tiny}{Relative size of tiny text (e.g., caption)}

\item{rel_large}{Relative size of large text (e.g., title)}
}
\description{
Same with theme_cowplot except without xy axes and xy ticks.
}
