% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_slopegraph.r
\name{theme_slopegraph}
\alias{theme_slopegraph}
\title{A ggplot2 theme for plotting slopegraphs}
\usage{
theme_slopegraph(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Number, base font size}

\item{base_family}{String, base font family}
}
\value{
ggplot2 theme
}
\description{
A ggplot2 theme for plotting slopegraphs
}

