% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack_cols}
\alias{unpack_cols}
\title{Unpack cols argument to agg_data}
\usage{
unpack_cols(x)
}
\arguments{
\item{x}{cols}
}
\value{
list
}
\description{
Unpack cols argument to agg_data
}
\examples{
\dontrun{
  unpack_cols("x")
  unpack_cols(c(x = "x", y = "x + y", expr = "~ x + y"))
}
}
