% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labs.R
\name{add_labs}
\alias{add_labs}
\title{Add x, y labels, title, subtitle and caption to a plot.}
\usage{
add_labs(p, xlab, ylab, title, subtitle, caption)
}
\arguments{
\item{p}{: a ggplot object.}

\item{xlab}{: string, the x-axis label.}

\item{ylab}{: string, the y-axis label.}

\item{title}{: string, title of the plot. You should use the title to
concisely state the main insight the plot reveals. Avoid
using generic statement as plot title. For example,
"Income vs. Employment Length" is a bad title because
it just re-state what the y and x axes are.}

\item{subtitle:}{string, subtitle of the plot. You can use the subtitle
to explain the details of the insight. You can even write
one or more short paragraphs if you have to.}

\item{caption:}{string, caption of the plot. You can put anything else
useful in caption. For example, data source, method referece,
url and etc.}
}
\value{

}
\description{
\code{add_labs} takes a ggplot object as input and adds to it user-supplied
x, y labels, plot title, subtitle and caption. If user doesn't provide
values to these parameters, it'll just use the values from the input plot
without making changes.
}
\examples{
inst/examples/ex-add_labs.R
}
