% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ads}
\alias{ads}
\title{Internet Ads Revenue}
\format{
A data frame with 15 rows and 13 variables:
}
\usage{
ads
}
\description{
A dataset containing internet ads revenue from 2000 to 2014.
Source: iab.net
}
\keyword{datasets}
