% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\title{Aggregates data}
\usage{
agg_data(data, cols = names(data), group_by = NULL, agg_fun = function(x)
  sum(x, na.rm = TRUE), group_by2 = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Named character vector.}

\item{group_by}{A vector of column names.}

\item{agg_fun}{Function to use for aggregating.}
}
\value{
An aggregated data.frame.
}
\description{
Aggregates data
}
\examples{
library(ukbabynames)
agg_data(ukbabynames, c(Sex = "sex", Total = "n"))
agg_data(ukbabynames["sex"])
agg_data(ukbabynames[c("sex", "n")])
agg_data(ukbabynames, "n", "year")
agg_data(ukbabynames, "n", c("sex", "year"))
agg_data(ukbabynames, c(x = "year", y = "n"), c(x = "year"))
}
