% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\title{Aggregates data}
\usage{
agg_data(data, cols = names(data), group_by = not_numeric(x),
  agg_fun = function(x) sum(x, na.rm = TRUE))
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Named character vector.}
}
\value{
data.frame
}
\description{
Aggregates data
}
\examples{
agg_data(fruit, c(Product = "Product", y = "Units"))
}
