% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_plot.R
\name{area_plot}
\alias{area_plot}
\title{area_plot}
\usage{
area_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  size = 12, reorder = c("group", "facet_x", "facet_y"), palette = ez_col,
  ylabels = if (position == "fill") {     function(x) ez_labels(100 * x,
  append = "\%") } else {     ez_labels }, use_theme = theme_ez,
  position = c("stack", "fill"), facet_scales = "fixed")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character. Evaluates to a column.}

\item{y}{A named character.Evaluates to a column.}

\item{group}{character. Evaluates to a column.}

\item{facet_x}{character. Evaluates to a column.}

\item{facet_y}{character. Evaluates to a column.}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{reorder}{A character vector specifying the group variables to reorder.
Default is \code{c("group", "facet_x", "facet_y")}.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}

\item{position}{Default is \code{"stack"}}

\item{facet_scales}{Option passed to scales argument in \code{facet_wrap} or
\code{facet_grid}. Default is \code{"fixed"}.}
}
\value{
A ggplot object.
}
\description{
Aggregates a data.frame and creates a stacked area chart.
}
\examples{
area_plot(mtcars, "carb", "1", size = 12)
area_plot(mtcars, "carb", "1", "cyl", use_theme = ggplot2::theme_bw)
area_plot(mtcars, "carb", "1", "cyl", reorder = NULL)
area_plot(mtcars, "cyl", "1", "carb", position = "fill")
area_plot(mtcars, "carb", c(Count = "1"), size = 12)
area_plot(mtcars, "carb", c(Count = "1"), "cyl", "gear")
area_plot(mtcars, "carb", "1", "cyl", "gear", "am", position = "fill")
}
