% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attacks_by_country}
\alias{attacks_by_country}
\title{Internet attacks data by each country}
\format{
A data frame with 1680 rows and 5 columns:
}
\source{
\url{https://github.com/hrbrmstr/facetedcountryheatmaps/tree/master/data}
}
\usage{
attacks_by_country
}
\description{
A dataset containing logs of internet attacks by country for each hour
of a day during a fixed period. It's derived from this raw data source:
https://github.com/hrbrmstr/facetedcountryheatmaps/tree/master/data
}
\keyword{datasets}
