% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{bar_plot}
\usage{
bar_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  size = 20, width = NULL, reorder = c("group", "facet_x", "facet_y"),
  palette = ez_col, ylabels = if (position == "fill") {     function(x)
  ez_labels(100 * x, append = "\%") } else {     ez_labels },
  use_theme = theme_ez, position = "stacked")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{group}{character}

\item{facet_x}{character}

\item{facet_y}{character}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}
}
\value{
A ggplot object.
}
\description{
bar_plot
}
\examples{
bar_plot(mtcars, "carb", "1", size = 12)
bar_plot(mtcars, "carb", "1", "cyl", use_theme = ggplot2::theme_bw)
bar_plot(mtcars, "carb", "1", "cyl", reorder = NULL)
bar_plot(mtcars, "carb", "1", "cyl", reorder = NULL, position = "fill")
bar_plot(mtcars, "carb", c(Count = "1"), size = 12)
bar_plot(mtcars, "carb", c(Count = "1"), "cyl", "gear")
bar_plot(mtcars, "carb", "1", "cyl", "gear", "am", position = "fill")
}
