% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{bar_plot}
\usage{
bar_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  size = 12, width = NULL, reorder = c("group", "facet_x", "facet_y"),
  palette = ez_col, ylabels = if (position == "fill") {     function(x)
  ez_labels(100 * x, append = "\%") } else {     ez_labels }, y_rescale = 1,
  label_cutoff = 0.12, use_theme = theme_ez, position = "stack",
  facet_scales = "fixed")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{group}{character}

\item{facet_x}{character}

\item{facet_y}{character}

\item{size}{theme size for \code{use_theme()}. Default is 12.}

\item{width}{Width of bar.}

\item{reorder}{Vector of dimensions to order by aggregation of y. Default is
\code{c("group", "facet_x", "facet_y")}}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{y_rescale}{Rescaling factor for y-axis limits}

\item{label_cutoff}{Cutoff size (proportion of limit range) for excluding labels}

\item{use_theme}{ggplot theme function}

\item{position}{Either \code{"stack"} (default) or \code{"fill"}}

\item{facet_scales}{Scales argument passed to \code{fact_wrap()} or \code{facet_grid()}}
}
\value{
A ggplot object.
}
\description{
bar_plot
}
\examples{
df = ez_data()
bar_plot(df, "year", "units")
bar_plot(df, "year", "units", "fct", use_theme = ggplot2::theme_bw)
bar_plot(df, "year", "units", "fct", reorder = NULL)
bar_plot(df, "year", "units", "fct", reorder = NULL, position = "fill")
bar_plot(df, "year", c(Units = "units"))
bar_plot(df, "year", c(Units = "units"), "fct", "char")
bar_plot(df, "year", "units", "fct", "char", "num", position = "fill")
bar_plot(df, "year", "ifelse(fct == 'X', units, -units)", "fct")
}
