% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bo_bt_ratio_by_year}
\alias{bo_bt_ratio_by_year}
\title{Boxoffice/Budget Ratio}
\format{A data frame with 101 rows and 2 variables:
\describe{
     \item{year}{Year of release.}
     \item{bo_bt_ratio}{Boxoffice/Budget Ratio}
}}
\usage{
bo_bt_ratio_by_year

bo_bt_ratio_by_year
}
\description{
A dataset containing the Boxoffices/Budget ratio from 1913 to 2014 of 
5944 films obtained from IMDB.com.

A dataset containing the Boxoffices/Budget ratio from 1913 to 2014 of 
5944 films obtained from IMDB.com.
}
\keyword{datasets}

