% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cb_color.R
\name{cb_color}
\alias{cb_color}
\title{Colorblind-friendly colors}
\usage{
cb_color(name)
}
\arguments{
\item{name}{The name of a color or palette. Possible values are 'black',
       "gray", "grey", "orange", "sky_blue", "bluish_green", "yellow",
       "blue", "vermilion", "reddish_purple", "cb_gray", "cb_grey", and
       "cb_black"}
}
\value{
A vector of hex strings
}
\description{
\code{cb_color} looks up the hex string of 9 colorblind-friendly colors given
by \url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/.}

General advice: for thin lines and small objects, darker blue and orange
are preferable to sky blue and yellow. When combining colors,
Use "warm" and "cool" colors alternatively. When using two warm colors or
two cool colors, put distinct differences in brightness or saturation.
Avoid combination of colors with low saturation or low brightness.
}
\examples{
# get single color
cb_color("black")
cb_color("gray")
cb_color("grey")
cb_color("orange")
cb_color("sky_blue")
cb_color("bluish_green")
cb_color("yellow")
cb_color("blue")
cb_color("vermilion") # red
cb_color("reddish_purple")

# get a palette of 8 colors with gray
cb_color("cb_gray")
cb_color("cb_grey")

# get a palette of 8 colors with black
cb_color("cb_black")
}

