% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params_expdist.R
\name{est_params_expdist}
\alias{est_params_expdist}
\title{Estimate the rate of the exponential distribution from which the
sample data were taken.}
\usage{
est_params_expdist(df, varname, digits = 4)
}
\arguments{
\item{df}{A data frame.}

\item{varname.}{String, name of a continuous variable with an empirical
distribution approximately exponential.}
}
\value{
An estimate of the rate of a theoretical exponential distribution.
}
\description{
Given a data frame and the name of a continuous variable from the data frame,
suppose the CCDF of the variable on a log-y scale is linear, then its
distribution can be modeled by an exponential distribution, of which the rate
can be estimated by this function.
}
\examples{
library(ezplot)
est_params_expdist(births, 'diffs')
est_params_expdist(births, 'diffs', digits = 5)
}
