% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezplot.R
\name{ezplot}
\alias{ezplot}
\title{Easy plotting of ggplot2 objects}
\usage{
ezplot(data, plot_type = "bar", x, y, group = NULL,
  group_method = "stack", facet = NULL, palette = NULL)
}
\arguments{
\item{data}{(\code{data.frame} or \code{tibble}) Data to be used in plot}

\item{plot_type}{\code{character(1)} Type of plot to create. Default is \code{'bar'}
\itemize{
  \item \code{'bar'} Create bar plot using \code{geom_bar()}
  \item \code{'line'} Create line plot using \code{geom_line()}
  \item \code{'point'} Create point plot (scatterplot) using \code{geom_point()}
  \item \code{'area'} Create area plot using \code{geom_area()}
}}

\item{x}{Variable to use on X-axis. Will be coerced to a factor}

\item{y}{Variable to use on Y-axis}

\item{group}{(Optional) Variable to group by}

\item{group_method}{\code{character(1)} How to display grouping. Default is \code{'stack'}
\itemize{
  \item \code{'stack'} places groups on top of one another, i.e. stacked bar chart
  \item \code{'dodge'} places groups side-by-side, i.e. clustered bar chart or butted bar chart
  \item \code{'fill'} places groups on top of one another and normalizes to
  the same height. Useful for comparing proportions
}}

\item{facet}{(Optional) Variable to facet by}

\item{fill}{(Optional) Variable to fill by}

\item{color}{(Optional) Variable to color by}
}
\description{
Creates a ggplot2 object using greatly simplified syntax
}
