% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{format_as_pct}
\alias{format_as_pct}
\title{Format numbers as percent.}
\usage{
format_as_pct(x, digits = 4)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{An integer indicating the number of decimal places to be used.}
}
\value{
A character vector that's just x in percent format.
}
\description{
\code{format_as_pct} takes a numeric argument and returns a
character vector of the original numeric vector in percent format.
}
\examples{
a = c(0.84984554, 0.18, 0.9273, 0.128294)
format_as_pct(a)
}

