% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bar_labels_freq.R
\name{get_bar_labels_freq}
\alias{get_bar_labels_freq}
\title{Helper function for \code{barplot_freq}. Not for external use.}
\usage{
get_bar_labels_freq(df, xvar, fillby, show_pct_on_y)
}
\arguments{
\item{df}{data frame of the original data where bar charts are made of.}

\item{xvar}{string, name of x categorical variable.}

\item{fillby}{string, name of a second categorical variable used for
sub-dividing the count or percent and coloring the bars.}

\item{show_pct_on_y}{logical, if TRUE, show percent on y-axis. Otherwise,
show count. Default is FALSE.}
}
\value{
a data frame containing bar labels and their positions
}
\description{
\code{get_bar_labels_freq} takes in the original data frame, the
x and fillby variable. It returns a data frame of labels and their postions
for \code{barplot_freq}.
}
\seealso{
\code{\link{mk_barplot_freq}}.
}
