% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{line_plot}
\usage{
line_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL)
}
\arguments{
\item{data}{data.frame}

\item{x}{named value}

\item{y}{named value}

\item{group}{value}

\item{facet_x}{value}

\item{facet_y}{value}
}
\description{
Creates line plots.
}
\examples{
line_plot(fruit, "OBS", "Units")
line_plot(fruit, "OBS", "Units", "Product")
line_plot(fruit, "OBS", "Units", "Product", "Size")
line_plot(fruit, "OBS", "Units", "Product", "Size", "Store")
}
