% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{line_plot}
\usage{
line_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  linesize = 1, size = 12, palette = ez_col, ylabels = ez_labels,
  use_theme = theme_ez, facet_scales = "fixed")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{group}{character}

\item{facet_x}{character}

\item{facet_y}{character}

\item{linesize}{width of line for \code{geom_line()}. Default is 1.}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}
}
\value{
A ggplot object.
}
\description{
Creates line plots.
}
\examples{
line_plot(mtcars, "cyl", "1", use_theme = ggplot2::theme_bw)
line_plot(mtcars, "cyl", c(Count = "1"))
line_plot(mtcars, "cyl", c(Count = "1"), "gear")
line_plot(mtcars, "cyl", c(Count = "1"), "gear", "am", size = 12)
line_plot(mtcars, "cyl", c(Count = "1"), "vs", "gear", "am", size = 12)
line_plot(mtcars, "cyl", c(Count = "1"), "gear", "am", facet_scales = "free_y")
}
