% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{line_plot}
\usage{
line_plot(data, x, y, group = NULL, facet_x = NULL, facet_y = NULL,
  linesize = 1, size = 12, palette = ez_col, ylabels = ez_labels,
  use_theme = theme_ez, facet_scales = "fixed")
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character. Evaluates to a column.}

\item{y}{A named character. Evaluates to a column.}

\item{group}{character. Evaluates to a column.}

\item{facet_x}{character. Evaluates to a column.}

\item{facet_y}{character. Evaluates to a column.}

\item{linesize}{width of line for \code{geom_line()}. Default is 1.}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{ylabels}{label formatting function}

\item{use_theme}{ggplot theme function}

\item{facet_scales}{Option passed to scales argument in \code{facet_wrap} or
\code{facet_grid}. Default is \code{"fixed"}.}
}
\value{
A ggplot object.
}
\description{
Creates line plots.
}
\examples{

df = ez_data()
line_plot(ez_data(), "week", "value", use_theme = ggplot2::theme_bw)
line_plot(ez_data(), "week", c("Sales ($)" = "value"))
line_plot(ez_data(), "week", "value", "char")
line_plot(ez_data(), "week", "value", "char", "fct")
line_plot(ez_data(), "week", "value", "char", "fct", "num", facet_scales = "free_y")
line_plot(ez_data(), "year2", "~ value / units", "char", "fct", "num")

}
