% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mk_barplot.R
\name{mk_barplot}
\alias{mk_barplot}
\title{Create a function that draws ggplot2 barplots}
\usage{
mk_barplot(df)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
\code{function(xvar, yvar, fillby, xlab="", ylab="", main="", ypct=F, ypct_jump=0.2, legend=T)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item fillby   :  string, the variable used for coloring the bars.
     \item xlab     :  string, the x-axis label.
     \item ylab     :  string, the y-axis label.
     \item main     :  string, the title of the plot.
     \item ypct     :  logical, indicating whether to use percent format on y-axis. Default is FALSE.
     \item ypct_jump:  numeric, between 0 and 1. Default is 0.1
     \item legend   :  logical, indicating whether to show the legend. Default is TRUE.
}
}
\description{
\code{mk_barplot()} takes a data frame as input and returns a function that
can be used to make barplots on variables in the data frame.
}
\examples{
# make some fake data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
barplt("student", "grade", "student")
barplt("student", "grade", "student", legend=F)

# make some fake data
df2 = read.table(header=TRUE, text='
group level val
A      small 1.8
A      medium 2.2
A      large 1.5
B      small 2.0
B      medium 2.6
B      large 1.0
C      small 2.5
C      medium 1.3
C      large 2.9')

# draw a barplot
barplt = mk_barplot(df2)
barplt("group", "val", "level")

# calculate the percentage of the levels within each group
library(tidyr)
library(dplyr)
pct = df2 \%>\% spread(level, val)
temp = pct[, 2:4]
pct = cbind(group=pct[, 1], temp / apply(temp, 1, sum))
pct = pct \%>\% gather(level, pct, -group)
pct$level = factor(pct$level, levels=c("small", "medium", "large"))

# plot a stacked barplot to display the percentages
barplt = mk_barplot(pct)
barplt("group", "pct", "level", ypct=T)
}

