% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barplot_tufte.R
\name{mk_barplot_tufte}
\alias{mk_barplot_tufte}
\title{Create a function that draws tufte bar charts.}
\usage{
mk_barplot_tufte(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, xorder="alphanumeric", bar_w=0.25, font_size=14, 
               font_family="serif", show_axis_ticks=F, ybreak_n=5, ybreak_h=1, 
               ytick_decimals=0, main="")}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item xorder   :  string, takes on one of three values: "alphanumeric", 
                       "ascend", or "descend". It arranges the bars by
                       their heights accordingly.
     \item bar_w    :  numeric, width of the bars. Default = 0.25.
     \item font_size:  numeric, size of text. Default = 14.
     \item font_family:  string, default = "serif".
     \item show_axis_ticks:  logical, display axis ticks if TRUE. 
     \item ybreak_n  :  integer, number of y labels. We'll break the bars at 
                        into chunks at these positions. Default = 5.
     \item ybreak_h  :  numeric, the height of each break. Default = 1.
     \item ytick_decimals  :  integer, the number of decimal places of each
                              y tick text shown. Default = 0.
     \item main  :  string, title of the plot.                         
}
}
\description{
Inspired and modified after Lukasz Piwek's article at 
http://motioninsocial.com/tufte/#minimal-barchart
}

