% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barploth_resp.R
\name{mk_barploth_resp}
\alias{mk_barploth_resp}
\title{Create a function for making publishable ggplot2 horizontal barplots,
showing single or cumulative values of a continuous variable by 1 or 2
categorical variables.}
\usage{
mk_barploth_resp(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby = "1", yorder = "alphanumeric",
               show_pct = FALSE, label_decimals = 2, label_size = 3,
               font_size = 14)}
\itemize{
     \item xvar     :  string, name of a continuous variable for x-axis.
     \item yvar     :  string, name of a categorical variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       subdividing and coloring the bars. Default = "1",
                       meaning no such variable is supplied.
     \item yorder   :  string, "alphanumeric", "ascend" or "descend". It
                       specifies how categories are ordered on the y-axis.
                       Default = "alphanumeric".
     \item show_pct :  logical, if TRUE, format x-axis and bar labels as %;
                       otherwise, format them as comma. Default is FALSE.
     \item label_decimals: integer, the number of decimal points shown
                       on the bar labels. Default = 2.
     \item label_size: integer, size of bar label text. Default = 3.
                       Hide bar labels when its value is 0.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_barploth_resp} takes a data frame as input and returns a function
for making bar charts with any categorical variable from the data frame on
the y-axis and the value (or cumulative values) of any continuous variable on
the x-axis. The output function can also produce dodged bar charts when
supplied a second categorical variable, a fillby variable. The resulting
bar chart will have bars ordered ordered by the alphanumerical order
of the y levels by default, and labeled with the cumulative x values.
If the x values are decimals, user can specify the number of
decimals to display on the bar labels. If the x values are between 0 and 1,
user can choose to format the x-axis and the bar labels as percent (%). Plus,
the resulting plot will have a clean theme and clear fonts.
It'll also use color-blind friendly palettes.
}
\examples{
library(ezplot)

g = mk_barploth_resp(films)
p = g("boxoffice", "mpaa", yorder = "descend", font_size = 10)
add_labs(p, title = "Fuel efficiency generally decreases with engine size",
         subtitle = "Two seaters (sports cars) are an exception ...",
         caption = "Data from fueleconomy.gov")
# use label_size = 0 to remove labels
g("bo_bt_ratio", "mpaa", fillby = "year_cat", label_decimals = 1, label_size = 0)

library(dplyr)
df = films \%>\% count(mpaa, made_money) \%>\% mutate(pct = n / sum(n))
g = mk_barploth_resp(df)
g("pct", "mpaa")
g("pct", "mpaa", label_decimals = 1, yorder = "descend")
g("pct", "mpaa", show_pct = T, label_decimals = 3, font_size = 9) \%>\%
   add_labs(ylab = NULL,
            title = "There're more R rated films than NC-17, PG and PG-13 combined.",
            subtitle = "Although a substantial portion of the films have mpaa rating info missing.",
            caption = "data were scrapped from imdb.com in 2010")
g("pct", "mpaa", fillby = "made_money", show_pct = T)
}
\seealso{
\code{\link{scale_axis}} for adding different scales to the y-axis.
}
