% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_boxplot.R
\name{mk_boxplot}
\alias{mk_boxplot}
\title{Create a function that draws ggplot2 boxplots}
\usage{
mk_boxplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, xlab="", ylab="", main="", vpos=0, legend=TRUE)}
\itemize{
     \item xvar     :  string, x variable.
     \item yvar     :  string, y variable.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item lab_at_top: logical, if T, label the boxplots at the top. Default is TRUE.
     \item main     :  string, title of the plot.
     \item vpos     :  number, additional height of the text labels beyond the max y-value of each group.
     \item legend   :  logical, show legend or not. Default = TRUE.
}
}
\description{
\code{mk_boxplot} takes a data frame as input and returns a function that
can be used to make boxplots using variables in the data frame.
}
\examples{
library(ezplot)
plt = mk_boxplot(films)

# plot distributions of budget over the years
title1 = "Annual Distribution of Budget from 1913 to 2014"
p = plt("year", "budget", ylab="budget", main=title1)
p = scale_axis(p, "y", scale="comma")
print(p)
scale_axis(p, scale="log10")

# plot distributions of boxoffice over the years
title2 = "Annual Distribution of Boxoffice from 1913 to 2014"
p = plt("year", "boxoffice", ylab="boxoffice", main=title2)
start = min(films$year)
end = max(films$year)
p = scale_axis(p, "y", scale="log10")
p = p + ggplot2::scale_x_continuous(limits = c(start, end),
                                    breaks = seq(start, end, 10))
print(p)

# plot distributions of boxoffice at each aggregated year level
p = plt("year_cat", "boxoffice", ylab="boxoffice")
p = scale_axis(p, scale="log10")
print(p)

# plot distributions of budget at each aggregated year level
p = plt("year_cat", "budget", ylab="boxoffice", legend=F)
scale_axis(p, scale="log10")
}

