% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_boxplot.R
\name{mk_boxplot}
\alias{mk_boxplot}
<<<<<<< HEAD
\title{Create a function that draws ggplot2 boxplots}
=======
\title{Create a function for making publishable ggplot2 boxplots.}
>>>>>>> v1.0.0
\usage{
mk_boxplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
<<<<<<< HEAD
\code{function(xvar, yvar, fillby="1", xlab="", ylab="", main="", legend=TRUE, 
               add_label=TRUE, lab_at_top=TRUE, vpos=0)}
\itemize{
     \item xvar     :  string, x variable.
     \item yvar     :  string, y variable.
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot. 
     \item legend   :  logical, show legend or not. Default = TRUE.
     \item add_label:  logical, whether or not to put number of observations as labels. Default is TRUE.
     \item lab_at_top: logical, whether to put the labels at the top or bottom. Default is TRUE.
     \item vpos     :  number, additional height of the text labels beyond the max y-value of each group.
}
}
\description{
\code{mk_boxplot} takes a data frame as input and returns a function that 
can be used to make boxplots using variables in the data frame.
}
\examples{
library(ezplot)
plt = mk_boxplot(films)

# plot distributions of budget over the years
title1 = "Annual Distribution of Budget from 1913 to 2014"
p = plt("year", "budget", ylab="budget", main=title1)
p = scale_axis(p, "y", scale="comma")
print(p)
scale_axis(p, scale="log10")

# plot distributions of boxoffice over the years
title2 = "Annual Distribution of Boxoffice from 1913 to 2014"
p = plt("year", "boxoffice", ylab="boxoffice", main=title2)
start = min(films$year)
end = max(films$year)
p = scale_axis(p, "y", scale="log10")
p = p + ggplot2::scale_x_continuous(limits = c(start, end), 
                                    breaks = seq(start, end, 10))
print(p)

# plot distributions of boxoffice at each aggregated year level
p = plt("year_cat", "boxoffice", ylab="boxoffice")
p = scale_axis(p, scale="log10")
print(p)

# plot distributions of budget at each aggregated year level
p = plt("year_cat", "budget", ylab="boxoffice", legend=F)
scale_axis(p, scale="log10")
=======
\code{function(xvar, yvar, fillby = "1", notched = FALSE, label_size = 3,
               font_size = 14)}
\itemize{
     \item xvar     :  string, name of a categorical variable for x-axis.
     \item yvar     :  string, name of a continuous variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       breaking down the y values of each box. Default = "1",
                       meaning no such variable is supplied.
     \item notched  :  logical, draw notched boxplots when TRUE;
                       otherwise, draw regular boxplots. Default = FALSE.
     \item label_size: integer, size of bar label text. Default = 3.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_boxplot} takes a data frame as input and returns a function for
making boxplots with any categorical variable from the data frame on the
x-axis and any continuous variable on the y-axis. The output function can
also produce dodged boxplots when supplied a second categorical variable, a
fillby variable.
}
\examples{
inst/examples/ex-mk_boxplot.R
>>>>>>> v1.0.0
}
