% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_boxplot.R
\name{mk_boxplot}
\alias{mk_boxplot}
\title{Create a function for making publishable ggplot2 boxplots.}
\usage{
mk_boxplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar = "1", yvar, fillby = "1", notched = FALSE,
               legend_title = fillby, legend_pos = "right",
               label_size = 3, font_size = 14)}
\itemize{
     \item xvar     : string, name of a categorical variable for x-axis.
                      Default = "1", just draw a boxplot of yvar by itself.
     \item yvar     :  string, name of a continuous variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       breaking down the y values of each box. Default = "1",
                       meaning no such variable is supplied.
     \item notched  :  logical, draw notched boxplots when TRUE;
                       otherwise, draw regular boxplots. Default = FALSE.
     \item legend_title: string, legend title. Default is the name of the
                         fillby variable.
     \item legend_pos:   string, legend position. Default = "right".
     \item label_size: integer, size of bar label text. Default = 3.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_boxplot} takes a data frame as input and returns a function for
making boxplots with any categorical variable from the data frame on the
x-axis and any continuous variable on the y-axis. The output function can
also produce dodged boxplots when supplied a second categorical variable, a
fillby variable.
}
\examples{
inst/examples/ex-mk_boxplot.R
}
