% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_densityplot.R
\name{mk_densityplot}
\alias{mk_densityplot}
\title{Create a function for making publishable ggplot2 density plots to
show distributions over time/space or within different groups.}
\usage{
mk_densityplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar = "1", cut_tail = 0.005, font_size = 14)}
\itemize{
     \item xvar : string, name of a continuous variable for x-axis.
     \item yvar : string, name of a categorical variable for y-axis. It's
                  used to group x. Default is "1", meaning no groups.
     \item cut_tail : a number between 0 and 1. Lines with heights below this
                      cutoff will be removed. The cutoff is relative to the
                      overall maximum, so cut_tail = 0.01 would remove
                      everything that is 1% or less than the highest point.
                      Default = 0.005, so 0.5% or below are removed.
     \item font_size: overall font size. Default = 14. The font size of the
                      axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_densityplot} takes a data frame as input and returns a function for
making density plots of any continuous variable (x) from the data frame for
each group of a categorical variable (y).
}
\examples{
inst/examples/ex-mk_densityplot.R
}
