% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_distplot.R
\name{mk_distplot}
\alias{mk_distplot}
\title{Create a function that draws ggplot2 histograms or density plots.}
\usage{
mk_distplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, fillby="", xlab="", type="histogram", binw=NULL, main="",
               add_vline_mean=FALSE, add_vline_median=FALSE)}
\itemize{
     \item xvar     :  string, x variable.
     \item fillby   :  string, grouping variable. Default = "".
     \item xlab     :  stirng, x-axis label.
     \item type     :  string, "density" or "histogram". Default is "histogram".
     \item binw     :  number, bin width. Default = NULL.
     \item main     :  string, title of the plot.
     \item add_vline_mean: logical, whether to add a vertical line at the mean. Default=FALSE.
     \item add_vline_median: logical, whether to add a vertical line at the median. Default=FALSE.
}
}
\description{
\code{mk_distplot} takes a data frame as input and returns a function that can 
be used to make histograms or density plots using variables in the data frame.
It uses \code{draw} as a helper.
}
\examples{
f = mk_distplot(iris)
f("Sepal.Length")
f("Sepal.Length", binw=0.3)
f("Sepal.Length", binw=0.3, add_vline_mean=T)
f("Sepal.Length", binw=0.3, add_vline_median=T)
f("Sepal.Length", binw=0.3, add_vline_mean=T, add_vline_median=T)

f("Sepal.Length", type="density")
f("Sepal.Length", type="density", add_vline_mean=T)
f("Sepal.Length", type="density", add_vline_median=T)
f("Sepal.Length", type="density", add_vline_median=T, add_vline_mean=T)

f("Sepal.Length", fillby="Species")
f("Sepal.Length", fillby="Species", binw=0.3)
f("Sepal.Length", fillby="Species", binw=0.3, add_vline_mean=T)
f("Sepal.Length", fillby="Species", binw=0.3, add_vline_median=T)
f("Sepal.Length", fillby="Species", type="density")
f("Sepal.Length", fillby="Species", type="density", add_vline_mean=T)
f("Sepal.Length", fillby="Species", type="density", add_vline_median=T)
}
\seealso{
\code{\link{draw}}.
}

