% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mk_distplot.R
\name{mk_distplot}
\alias{mk_distplot}
\title{Create a function that draws ggplot2 histograms or density plots.}
\usage{
mk_distplot(df)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
\code{function(xvar, fillby="", xlab="", type="histogram", binw=NULL, main="")}
\itemize{
     \item xvar     :  string, the x variable.
     \item fillby   :  string, the grouping variable. Default is "".
     \item xlab     :  stirng, the x-axis label.
     \item type     :  string, "density" or "histogram". Default is "histogram".
     \item binw     :  number, bin width. Default is NULL.
     \item main     :  string, the title of the plot.
}
}
\description{
\code{mk_distplot()} takes a data frame as input and returns a function that
can be used to make histograms or density plots on variables in the data frame.
It uses \code{draw()} as a helper.
}
\examples{
f = mk_distplot(iris)
f("Sepal.Length")
f("Sepal.Length", binw=0.3)
f("Sepal.Length", binw=0.3, add_vline_mean=T)
f("Sepal.Length", binw=0.3, add_vline_median=T)
f("Sepal.Length", binw=0.3, add_vline_mean=T, add_vline_median=T)

f("Sepal.Length", type="density")
f("Sepal.Length", type="density", add_vline_mean=T)
f("Sepal.Length", type="density", add_vline_median=T)
f("Sepal.Length", type="density", add_vline_median=T, add_vline_mean=T)

f("Sepal.Length", fillby="Species")
f("Sepal.Length", fillby="Species", binw=0.3)
f("Sepal.Length", fillby="Species", binw=0.3, add_vline_mean=T)
f("Sepal.Length", fillby="Species", binw=0.3, add_vline_median=T)
f("Sepal.Length", fillby="Species", type="density")
f("Sepal.Length", fillby="Species", type="density", add_vline_mean=T)
f("Sepal.Length", fillby="Species", type="density", add_vline_median=T)
}
\seealso{
\code{\link{draw}}.
}

