% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_heatmap.R
\name{mk_heatmap}
\alias{mk_heatmap}
\title{Create a function that draws ggplot2 heatmaps}
\usage{
mk_heatmap(df)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
\code{function(xvar, yvar, fillby, xlab="", ylab="", main="", base_size = 12,
               use_theme_gray=T, legend=T)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item fillby   :  string, the variable used for coloring the bars.
     \item xlab     :  string, the x-axis label.
     \item ylab     :  string, the y-axis label.
     \item main     :  string, the title of the plot.
     \item base_size:  numeric, the size of both axis tick text, default = 12.
     \item use_theme_gray:  logical, indicating whether to use theme_gray() or theme_minimal(). Default is TRUE.
     \item legend   :  logical, indicating whether to show the legend. Default is TRUE.
}
}
\description{
\code{mk_heatmap()} takes a data frame as input and returns a function that
can be used to make heatmaps on variables in the data frame.
}
\examples{
# make some fake data
}

