% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_heatmap.R
\name{mk_heatmap}
\alias{mk_heatmap}
\title{Create a function that draws ggplot2 heat maps.}
\usage{
mk_heatmap(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby, facet_by = NULL, facet_ncol = 1,
               palette = "C", color_direction = -1, legend_title = fillby,
               format_legend_as_comma = FALSE, font_size = 8)}
\itemize{
     \item xvar. String, name of a categorical variable for x-axis.
     \item yvar. String, name of another categorical variable for y-axis.
     \item fillby. String, name of a continuous variable, and its values are
     used to color the tiles.
     \item facet_by. String, name of a categorical variable for grouping and
     creating facets. Default = NULL.
     \item facet_ncol. Number of columns when facetting. Default = 1. Only
     works when facet_by is not NULL.
     \item palette. String, the colormap option to use. Possible values are
     "A", "B", "C" (default), "D" and "E".
     \item color_direction. 1 or -1. Sets the order of colors in the scale.
     If 1, colors are ordered from darkest to lightest. If -1 (default),
     ordered from lightest to darkest.
     \item legend_title. String, legend title. Default uses the name of the
     fillby variable.
     \item format_legend_as_comma. Logical. If TRUE, display numbers like
     2000 as 2,000 on legend. If FALSE (default), display numbers as they are.
     \item font_size. Overall font size. Default = 8. The font size of the
     axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_heatmap} takes a data frame as input and returns a function that
can be used to make heat maps using variables in the data frame.
}
\examples{
inst/examples/ex-mk_heatmap.R
}
