% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_histogram.R
\name{mk_histogram}
\alias{mk_histogram}
\title{Create a function for making publishable ggplot2 histograms.}
\usage{
mk_histogram(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, facet_by = NULL, binw = NULL, bins = 30, facet_ncol = 1,
               font_size = 14, add_vline_median = TRUE, add_vline_mean = TRUE)}
\itemize{
     \item xvar     : string, name of a continuous variable for x-axis.
     \item facet_by : string, name of a categorical variable for grouping
                      x and creating facets. Default = NULL.
     \item binw     : width of the bins. Can be specified as a number, or
                      a function that calculates width from x. Default = NULL.
     \item bins     : number of bins. Overridden by binw. Default = 30. You
                      almost always want to change the defult. Andrew Gelman
                      recommends to choose a sufficiently large number to
                      show how raw data are distributed without any smoothing
                      as happens when estimating density. Read this blog post
                      for details:
                      http://andrewgelman.com/2009/10/23/variations_on_t/
     \item facet_ncol: number of columns when facetting. Default = 1.
                     Only works when facet_by is not NULL.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
     \item add_vline_median: logical, if TRUE, add a vertical line at the
                             median of x. Default = TRUE.
     \item add_vline_mean: logical, if TRUE, add a vertical line at the mean
                           of x. Default = TRUE.
}
}
\description{
\code{mk_histogram} takes a data frame as input and returns a function for
making histograms of any continuous variable from the data frame. The output
function can also produce faceted histograms when supplied a categorical
variable, a facet_by variable, for grouping.
}
\examples{
inst/examples/ex-mk_histogram.R
}
\seealso{
\code{\link{get_summ_stats}}.
}
