% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lineplot.R
\name{mk_lineplot}
\alias{mk_lineplot}
\title{Create a function for making publishable ggplot2 line plots.}
\usage{
mk_lineplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby = "1", pt_size = 1, linew = 0.7,
               font_size = 14, xlab = xvar, ylab = yvar, ...)}
\itemize{
     \item xvar   : string, name of a continuous or categorical variable for
                    x-axis.
     \item yvar   : string, name of a continuous variable for y-axis.
     \item fillby : string, name of a categorical variable for
                    grouping and coloring the lines. Default = "1", meaning
                    no such variable is supplied.
     \item pt_size: number, size of the points. Default = 1.
     \item linew  : number, width of the line. Default = 0.7.
     \item font_size: overall font size. Default = 14. The font size of the
                      axes and legend text is a fraction of this value.
     \item xlab   : string, the x-axis label. Default is xvar.
     \item ylab   : string, the y-axis label. Default is yvar.
     \item ...    : other arguments for ggplot2::labs(), for example,
                    title, subtitle, caption and etc.
}
}
\description{
\code{mk_lineplot} takes a data frame as input and returns a function for
making lineplots with any continuous variables from the data frame on the y
axis, and any continuous or categorical variables from the data frame on the
x axis. The x variable often measures time, for example, years. When supplied
a categorical fillby variable, the output function will produce lineplots
where lines are colored differently according to the levels of the fillby
variable.
}
\examples{
library(ezplot)

# plot boxoffice/budget ratio over the years
plt = mk_lineplot(bo_bt_ratio_by_year)
title = "Boxoffice/Budget Ratio from 1913 to 2014"
p = plt("year", "bo_bt_ratio", xlab = NULL, ylab="boxoffice/budget ratio",
        title = title, caption = "data source: IMBD")
scale_axis(p, scale = "sqrt")

# plot total budget and boxoffice over the years
plt = mk_lineplot(btbo_by_year)
title = "Annual Total Budget and Boxoffice from 1913 to 2014"
p = plt("year", "tot", fillby = "type", font_size = 10, xlab = NULL,
        ylab = "total amount ($billion)", title = title)
scale_axis(p, scale = "log")

library(dplyr)
plt = mk_lineplot(films \%>\% group_by(year_cat) \%>\%
                  summarise(avg_boxoffice = mean(boxoffice)))
plt("year_cat", "avg_boxoffice", xlab = NULL, ylab = "Mean Boxoffice")
}
