% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mk_scatterplot.R
\name{mk_scatterplot}
\alias{mk_scatterplot}
\title{Create a function that draws ggplot2 scatterplots.}
\usage{
mk_scatterplot(df)
}
\arguments{
\item{df}{data frame containing variables to be visualized.}
}
\value{
\code{function(xvar, yvar, fillby="", xlab="", ylab="", main="",
               xlog=F, ylog=F, xlog10=F, ylog10=F,
               xpct=F, ypct=F, xpct_jump=0.2, ypct_jump=0.2,
               pt_alpha=NULL, pt_size=NULL)}
\itemize{
     \item xvar     :  string, the x variable.
     \item yvar     :  string, the y variable.
     \item fillby   :  string, the variable used for coloring the points
     \item xlab     :  string, the x-axis label.
     \item ylab     :  string, the y-axis label.
     \item main     :  string, the title of the plot.
     \item xlog     :  logical, indicating whether to use log scale on x-axis. Default is FALSE.
     \item ylog     :  logical, indicating whether to use log scale on y-axis. Default is FALSE.
     \item xlog10   :  logical, indicating whether to use log10 scale on x-axis. Default is FALSE.
     \item ylog10   :  logical, indicating whether to use log10 scale on y-axis. Default is FALSE.
     \item xpct     :  logical, indicating whether to use percent format on x-axis. Default is FALSE.
     \item ypct     :  logical, indicating whether to use percent format on y-axis. Default is FALSE.
     \item xpct_jump:  numeric, between 0 and 1. Default is 0.2
     \item ypct_jump:  numeric, between 0 and 1. Default is 0.2
     \item add_line :  logical, indicating whether to add a lm line. Default is F.
     \item linew    :  numeric, default is 1.
     \item pt_alpha :  numeric, specifying the transparency of the points. Default is NULL.
     \item pt_size  :  numeric, specifying the size of the points. Default is NULL.
}
}
\description{
\code{mk_scatterplot()} takes a data frame as input and returns a function that
can be used to make scatterplots on variables in the data frame.
}
\examples{
library(ezplot)
plt = mk_scatterplot(films)

plt("boxoffice", "budget", xlog10=T, ylog10=T)
plt("boxoffice", "budget", xlog10=T, ylog10=T, add_line=T)
plt("boxoffice", "budget", fillby="year_cat", xlog10=T, ylog10=T)
plt("boxoffice", "budget", fillby="year_cat", xlog10=T, ylog10=T, add_line=T)
plt("boxoffice", "budget", fillby="year_cat", xlog10=T, ylog10=T, add_line=T)
plt("boxoffice", "budget", fillby="made_money", xlog10=T, ylog10=T, add_line=T, linew=0.5)
}

