% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_scatterplot.R
\name{mk_scatterplot}
\alias{mk_scatterplot}
\title{Create a function for making publishable ggplot2 scatterplots.}
\usage{
mk_scatterplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby = "1", alpha = 0.8, pt_size = 1,
               jitter = FALSE, font_size = 14, add_cnt_to_legend = T)}
\itemize{
     \item xvar     :  string, name of a categorical variable for x-axis.
     \item yvar     :  string, name of a continuous variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       grouping and coloring the points. Default = "1",
                       meaning no such variable is supplied.
     \item alpha    :  a number between 0 and 1, transparency level of the
                       point colors. Smaller value means more transparent.
                       Default = 0.8.
     \item pt_size  :  number, size of the points. Default = 1.
     \item jitter   :  logical, jitter points if TRUE. Used when there're
                       overlapping points. Default = FALSE.
     \item font_size:  overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
     \item add_cnt_to_legend: logical, when TRUE (default), it will show
                   the number of non-missing records for each level in the
                   fillby var.
}
}
\description{
\code{mk_scatterplot} takes a data frame as input and returns a function for
making scatterplots with any continuous variables from the data frame on the
x and y axes. When supplied a categorical fillby variable, the output
function will produce scatterplots where the points are colored differently
according to the levels of the fillby variable.
}
\examples{
library(ezplot)
plt = mk_scatterplot(films)

plt("budget", "boxoffice")
plt("budget", "boxoffice") \%>\% add_labs(subtitle = "Total number of observations: 5,944")
plt("budget", "boxoffice", fillby = "year_cat", alpha = 0.2) \%>\%
    add_labs(title = "Boxoffice and Budget are related, try log scales",
             caption = "Source: IMDB")

plt("budget", "boxoffice", fillby = "year_cat", alpha = 0.2, add_cnt_to_legend = F) \%>\%
    add_labs(title = "Boxoffice and Budget are related, try log scales",
             caption = "Source: IMDB")

plt("year", "rating", font_size = 10) \%>\% add_labs(xlab = NULL)
plt("year", "rating", jitter = T) \%>\% add_labs(xlab = NULL)

df = data.frame(V1 = c(56, 123, 546, 26, 62, 6, NaN, NA, NA, 15),
                V2 = c(21, 231, 5, 5, 32, NA, 1, 231, 5, 200),
                V3 = c(NA, NA, 24, 51, 53, 231, NA, 153, 6, 700),
                V4 = c(2, 10, NA, 20, 56, 1, 1, 53, 40, 5000)
                )
nrow(df)

plt = mk_scatterplot(df)
plt("V1", "V2")
plt("V1", "V3")
}
