% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_scatterplot.R
\name{mk_scatterplot}
\alias{mk_scatterplot}
\title{Create a function that draws ggplot2 scatterplots.}
\usage{
mk_scatterplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby="", xlab="", ylab="", main="", add_line=FALSE,
               linew=1, pt_alpha=0.5, pt_size=1)}
\itemize{
     \item xvar     :  string, x variable.
     \item yvar     :  string, y variable.
     \item fillby   :  string, variable used for coloring the points
     \item xlab     :  string, x-axis label.
     \item ylab     :  string, y-axis label.
     \item main     :  string, title of the plot.
     \item add_line :  logical, whether to add a lm line. Default = FALSE.
     \item linew    :  number, default = 1.
     \item pt_alpha :  number, transparency of the points. Default = 0.5.
     \item pt_size  :  number, size of the points. Default = 1.
}
}
\description{
\code{mk_scatterplot} takes a data frame as input and returns a function that
can be used to make scatterplots using variables in the data frame.
}
\examples{
library(ezplot)
plt = mk_scatterplot(films)

plt("budget", "boxoffice", xlab="budget", ylab="boxoffice",
     main="Boxoffice vs. Budget")

p = plt("budget", "boxoffice", xlab="budget", ylab="boxoffice",
         pt_alpha=0.2, pt_size=1.5, add_line=T, linew=0.8)
scale_axis(p, "y", scale = "log")
scale_axis(p, "x", scale = "log")

p = scale_axis(p, "x", scale = "log10")
p = scale_axis(p, "y", scale = "log10")
print(p)

p = plt("budget", "boxoffice", fillby="year_cat", xlab="budget",
         ylab="boxoffice", add_line=T)
p = scale_axis(p, "x", scale = "log10")
p = scale_axis(p, "y", scale = "log10")
print(p)

p = plt("budget", "boxoffice", fillby="made_money", xlab="budget",
         ylab="boxoffice", add_line=T)
p = scale_axis(p, "x", scale = "log2")
p = scale_axis(p, "y", scale = "log2")
print(p)
}

