% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_scatterplot.R
\name{mk_scatterplot}
\alias{mk_scatterplot}
\title{Create a function for making publishable ggplot2 scatterplots.}
\usage{
mk_scatterplot(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(xvar, yvar, fillby = "1", alpha = 0.8, pt_size = 1,
               jitter = FALSE, legend_title = fillby, add_cnt_to_legend = TRUE,
               font_size = 14)}
\itemize{
     \item xvar     :  string, name of a categorical variable for x-axis.
     \item yvar     :  string, name of a continuous variable for y-axis.
     \item fillby   :  string, name of a different categorical variable for
                       grouping and coloring the points. Default = "1",
                       meaning no such variable is supplied.
     \item alpha    :  a number between 0 and 1, transparency level of the
                       point colors. Smaller value means more transparent.
                       Default = 0.8.
     \item pt_size  :  number, size of the points. Default = 1.
     \item jitter   :  logical, jitter points if TRUE. Used when there're
                       overlapping points. Default = FALSE.
     \item legend_title: string, legend title. Default uses the name of the
              fillby variable.
     \item add_cnt_to_legend: logical, when TRUE (default), it will show
                   the number of non-missing records for each level in the
                   fillby var.
     \item font_size:  overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
}
}
\description{
\code{mk_scatterplot} takes a data frame as input and returns a function for
making scatterplots with any continuous variables from the data frame on the
x and y axes. When supplied a categorical fillby variable, the output
function will produce scatterplots where the points are colored differently
according to the levels of the fillby variable.
}
\examples{
inst/examples/ex-mk_scatterplot.R
}
