% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{model_plot}
\usage{
model_plot(data, x = "ID", actual = "Actual", fitted = "Fitted",
  point_size = 2, size = 20)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A quoted expression.}

\item{actual}{A quoted expression.}

\item{fitted}{A quoted expression.}

\item{point_size}{Numeric. Default is 2.}

\item{size}{Theme base size. Default is 20.}
}
\value{
A ggplot object.
}
\description{
model_plot
}
\examples{
model_plot(data.frame(ID = 1:10, Actual = 1:10 + rnorm(10), Fitted = 1:10))
}
