% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA Player Performance Statistics}
\format{
A data frame with 50 rows and 21 columns:
}
\usage{
nba
}
\description{
A dataset containing 21 different performance statistics of 50 NBA players
obtained from flowingdata.com.
}
\keyword{datasets}
