% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf}
\alias{perf}
\title{perf}
\usage{
perf(fitted, actual, x_measure, y_measure)
}
\arguments{
\item{fitted}{Vector with values between 0 and 1}

\item{actual}{Vector with two levels}

\item{x_measure}{metric for ROCR::performance}

\item{y_measure}{metric for ROCR::performance}
}
\description{
Precision recall calculation
}
\examples{
ezplot:::perf(runif(1), sample(c(TRUE, FALSE), 1, replace = TRUE), "rpp", "lift")
ezplot:::perf(runif(10), sample(c(TRUE, FALSE), 10, replace = TRUE), "rpp", "lift")
ezplot:::perf(runif(5), sample(c(TRUE, FALSE), 5, replace = TRUE), "rec", "prec")
ezplot:::perf(runif(5), sample(c(TRUE, FALSE), 5, replace = TRUE), "fpr", "tpr")
}
