% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_plot.R
\name{pie_plot}
\alias{pie_plot}
\title{pie_plot}
\usage{
pie_plot(data, x, y = "1", facet_x = NULL, facet_y = NULL,
  ylabels = ez_labels, size = 12, label_cutoff = 0.04, round = Inf,
  signif = 3, palette = ez_col, reorder = c("x", "facet_x", "facet_y"),
  label_x = 0.8)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{Value.}

\item{y}{Value.}

\item{facet_x}{Value.}

\item{facet_y}{Value.}

\item{ylabels}{Label formatting function.}

\item{size}{Theme size for \code{use_theme()}. Default is 20.}

\item{label_cutoff}{Label cutoff value.}

\item{round}{Option for rounding label.}

\item{signif}{Option for retaining significant figures in label.}

\item{palette}{Colour function.}

\item{label_x}{Position of label from centre of pie.  0 is the centre of the
pie and 1 is the outer edge.}
}
\value{
ggplot object
}
\description{
pie_plot
}
\examples{
pie_plot(mtcars, "cyl", "1")
pie_plot(mtcars, "cyl", "1", reorder = NULL, label_x = 0.5)
pie_plot(mtcars, "cyl", "1", "gear", reorder = NULL, label_x = 0.5)
pie_plot(mtcars, "cyl", "1", "gear", "am")
pie_plot(mtcars, "cyl", "1", "gear", "am", reorder = NULL)
##pie_plot(mtcars, "cyl", "1", "am", "am")## WHY ERROR?
}
