% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plt_dist.R
\name{plt_dist}
\alias{plt_dist}
\title{Create a figure of 4 plots to show the distribution of a numeric variable.}
\usage{
plt_dist(dat, varname)
}
\arguments{
\item{dat}{a data frame or a matrix with colnames.}

\item{varname}{a string specifying the name of the variable of interest.}
}
\description{
\code{plt_dist()} takes in a data frame and a variable name, and plots the
histogram, boxplot, density curve and qqplot of the variable using the base
plot system.
}
\examples{
plt_dist(cars, "speed")
}

