% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_dist.R
\name{plt_dist}
\alias{plt_dist}
\title{Create a function that can be used to make a figure of 4 plots showing
the distribution of a continuous variable.}
\usage{
plt_dist(dat)
}
\arguments{
\item{dat}{A data frame or a matrix with colnames.}
}
\value{
\code{function(varname)}
\itemize{
     \item varname: string, the name of the continuous variable to be plotted
}
}
\description{
\code{plt_dist} takes in a data frame and returns a function that you can pass
in a variable name of any continuous variable of the data frame and it'll plot 
the histogram, boxplot, density curve and qqplot using the base plot system.
}
\examples{
f = plt_dist(cars)
f("speed")
f("dist")
}
