% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_plot.R
\name{prec_rec}
\alias{prec_rec}
\title{prec_rec}
\usage{
prec_rec(actual, fitted)
}
\arguments{
\item{actual}{Vector with two levels}

\item{fitted}{Vector with values between 0 and 1}
}
\description{
Precision recall calculation
}
\examples{
ezplot:::prec_rec(sample(c(TRUE, FALSE), 1, replace = TRUE), runif(1))
ezplot:::prec_rec(sample(c(TRUE, FALSE), 5, replace = TRUE), runif(5))
}
