% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_plot.R
\name{rel_plot}
\alias{rel_plot}
\title{rel_plot}
\usage{
rel_plot(data, x, y, group = NULL, size = 20, point_size = 2.5)
}
\description{
rel_plot
}
\examples{
rel_plot(fruit, "Units", "Value")
rel_plot(fruit, "Units", "Value", "paste(Product, Size, Store)")
rel_plot(fruit, "Product", "Value")
}
