% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_plot.R
\name{rel_plot}
\alias{rel_plot}
\title{relationship plot}
\usage{
rel_plot(data, x, y, group = NULL, size = 14, point_size = 2.5)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character value. Evaluates to a column.}

\item{y}{A named character value. Evaluates to a column.}

\item{group}{A character value. Evaluates to a column.}

\item{size}{theme size for \code{use_theme()}. Default is 14.}

\item{point_size}{Numeric. Default is 2.}
}
\description{
Examine the relationship between two vectors from data
}
\examples{
rel_plot(mtcars, "wt", "hp")
rel_plot(mtcars, "wt", "hp", "factor(cyl)")
rel_plot(mtcars, "factor(cyl)", "hp")
}
