% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc}
\alias{roc}
\title{roc}
\usage{
roc(actual, fitted)
}
\arguments{
\item{actual}{Vector with two levels}

\item{fitted}{Vector with values between 0 and 1}
}
\description{
Calculates ROC and AUC
}
\examples{
ezplot:::roc(sample(c(TRUE, FALSE), 1, replace = TRUE), runif(1))
ezplot:::roc(sample(c(TRUE, FALSE), 3, replace = TRUE), runif(3))
}
