% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{roc_plot}
\usage{
roc_plot(actual, fitted, size = 20)
}
\description{
roc_plot
}
\examples{
n = 500
actual = sample(c(FALSE, TRUE), n, replace = TRUE)
roc_plot(actual, actual)
roc_plot(actual, runif(n))
fitted = runif(n) ^ ifelse(actual == 1, 0.5, 2)
roc_plot(actual, fitted)

}
