% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rotate_axis_text.R
\name{rotate_axis_text}
\alias{rotate_axis_text}
\title{Rotate axis tick text}
\usage{
rotate_axis_text(p, text_angle_x = 45, text_angle_y = 0, hjust_x = NULL,
  hjust_y = NULL, vjust_x = NULL, vjust_y = NULL)
}
\arguments{
\item{p}{ggplot2 object.}

\item{axis_text_angle_x}{the angle of the tick text on x-axis. Default=45.}

\item{axis_text_angle_y}{the angle of the tick text on y-axis. Default=0.}
}
\value{
a ggplot2 object.
}
\description{
Rotate axis tick text
}
\examples{
# make some data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
p = barplt("student", "grade", "student", xlab="Student", ylab="Grade")
print(p)

# rotate the x-axis tick text
rotate_axis_text(p)
rotate_axis_text(p, 90)
rotate_axis_text(p, 90, 30)
}

