% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_axis_text.R
\name{rotate_axis_text}
\alias{rotate_axis_text}
\title{Rotate axis tick text}
\usage{
rotate_axis_text(p, text_angle_x = 45, text_angle_y = 0,
  hjust_x = NULL, hjust_y = NULL, vjust_x = NULL, vjust_y = NULL)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{text_angle_x}{Angle of tick text on x-axis. Default=45.}

\item{text_angle_y}{Angle of tick text on y-axis. Default=0.}

\item{hjust_x}{Vertical distance adjustment of the tick text to x-axis. Default=NULL.}

\item{hjust_y}{Horizontal distance adjustment of the tick text to y-axis. Default=NULL.}

\item{vjust_x}{Horizontal distance adjustment of the tick text to x-axis. Default=NULL.}

\item{vjust_y}{Vertical distance adjustment of the tick text to y-axis. Default=NULL.}
}
\value{
A ggplot2 object.
}
\description{
Rotate axis tick text
}
\examples{
# make some data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
p = barplt("student", "grade", "student", xlab="Student", ylab="Grade")
print(p)

# rotate the x-axis tick text
rotate_axis_text(p)
rotate_axis_text(p, 90)
rotate_axis_text(p, 90, 30, vjust_x=-2)
}
