% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpng.R
\name{save_png}
\alias{save_png}
\title{save_png}
\usage{
save_png(g, file, width, height, res, ...)
}
\arguments{
\item{g}{A ggplot or ezplot object.}

\item{file}{A path.}

\item{width}{Width of output image.}

\item{height}{Height or output image.}

\item{res}{Resolution of output image.}

\item{...}{Further arguments to pass to png}
}
\description{
save_png
}
