% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_axis.R
\name{scale_axis}
\alias{scale_axis}
\title{Use a different scale on either axis of a ggplot2 plot.}
\usage{
scale_axis(p, axis = "y", scale = "comma", pct_min = NULL,
  pct_max = NULL, pct_jump = NULL)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{axis}{A string of value "x" or "y". Default = "y".}

\item{scale}{A string of value "comma", "dollar", "pct", "log", "log1p",
"log10", "log2", "sqrt", or "exp".
 It specifies which scale to use. Default = "comma".}

\item{pct_min}{A number. The min value when using percent scale.
Default = 0.}

\item{pct_max}{A number. The max value when using percent scale. Default = 1.}

\item{pct_jump}{A number. The jump hight between ticks when using percent
scale. Default = 0.2.}
}
\value{
A ggplot2 object with the new scale applied to the input axis.
}
\description{
\code{scale_axis} takes a ggplot2 object as input and applies to its x-axis
or y-axis one of the following scales:
\itemize{
    \item comma  : show numbers in 000,000 format.
    \item dollar : show $ in front of numbers.
    \item pct    : use % format.
    \item log    : use log transformation.
    \item log1p  : use log(x+1) transformation.
    \item log10  : use log10 transformation.
    \item log2   : use log2 transformation.
    \item sqrt   : use squared root transformation.
    \item exp    : use exponential transformation.
}
It uses \code{scale_axis_helper} as a helper.
}
\examples{
library(ezplot)

#' # make some data
df = read.table(header=TRUE, text='
IQ score
125 0.86
110 0.72
108 0.64')

plt = mk_scatterplot(df)
p = plt("IQ", "score", pt_size = 2)
print(p)
scale_axis(p, scale = "pct")
scale_axis(p, scale = "pct", pct_max = 1, pct_min = 0, pct_jump = 0.2)


plt = mk_scatterplot(films)
p = plt("boxoffice", "budget", alpha=0.5, pt_size=1)
print(p)
scale_axis(p)
scale_axis(p, scale = "dollar")
p = scale_axis(p, scale = "log10")
print(p)
p = scale_axis(p, "x", scale = "log2")
print(p)

library(dplyr)
plt = mk_scatterplot(films \%>\% filter(bo_bt_ratio < 1))
p = plt("votes", "bo_bt_ratio", alpha=0.5, pt_size=1)
print(p)
scale_axis(p, scale = "sqrt")
scale_axis(p, scale = "exp")
scale_axis(p, "x", scale = "log")
scale_axis(p, "x", scale = "log2")
scale_axis(p, "x", scale = "log10")
scale_axis(p, "x", scale = "log1p")
}
\seealso{
\code{\link{scale_axis_helper}}.
}
