% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale_axis.R
\name{scale_axis}
\alias{scale_axis}
\title{Use a different scale on either axis of a ggplot2 plot.}
\usage{
scale_axis(p, axis = "y", scale = NULL, pct_max = 1, pct_jump = 0.2)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{axis}{A string of value "x" or "y". Default = "y".}

\item{scale}{A string of value "comma", "log", "log10", or "pct". It specifies which scale to use. Default = NULL, which uses the original scale.}

\item{pct_max}{A number no bigger than 1. It specifies the max value when using percent format. Default = 1.}

\item{pct_jump}{A number between 0 and 1. It specifies the interval width between ticks when using percent format. Default = 0.2.}
}
\value{
A ggplot2 object with the new scale applied to the input axis.
}
\description{
\code{scale_axis} takes a ggplot2 object as input and applies to its x-axis or
y-axis one of the following scales:
\itemize{
     \item comma: display big numbers in 000,000 format.
     \item log  : use log scale.
     \item log10: use log10 scale.
     \item pct  : use % format, the values must be between 0 and 1.
}
It uses \code{scale_axis_helper} as a helper.
}
\examples{
# make some fake data
library(ezplot)
plt = mk_scatterplot(films)

p = plt("boxoffice", "budget", pt_alpha=0.5, pt_size=1)
p = scale_axis(p, "y", scale = "log10")
print(p)

p = scale_axis(p, "x", scale = "log10")
print(p)

# make some data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')
df$pct = df$grade / sum(df$grade)

barplt = mk_barplot(df)
p = barplt("student", "pct", fillby="student", legend=F)
scale_axis(p, "y", scale="pct", pct_jump=0.3)
}
\seealso{
\code{\link{scale_axis_helper}}.
}

