% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_axis.R
\name{scale_axis_helper}
\alias{scale_axis_helper}
\title{Helper function used by \code{scale_axis}. Not for external use.}
\usage{
scale_axis_helper(lstr, axis, rstr)
}
\arguments{
\item{lstr}{string, ggplot2 code that comes before "x" or "y".}

\item{axis}{string, "x" or "y".}

\item{rstr}{string, ggplot2 code that comes after "x" or "y".}
}
\value{
a R expression to be evaluated by \code{scale_axis}.
}
\description{
\code{scale_axis_helper} takes in R code strings, combines them
and returns an expression to be evaluated by \code{scale_axis}.
}
\seealso{
\code{\link{scale_axis}}.
}
