% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_plot.R
\name{side_plot}
\alias{side_plot}
\title{side_plot}
\usage{
side_plot(data, x, y, ylabels = ez_labels, size = 12, palette = ez_col,
  signif = 3, reorder = TRUE, y_rescale = 1.25)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character.}

\item{y}{A named character.}

\item{ylabels}{label formatting function}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}
}
\description{
side_plot
}
\examples{
side_plot(mtcars, "gear", "1")
side_plot(mtcars, "cyl", c("Cars with <120 HP" = "hp < 120"))
side_plot(mtcars, "cyl", c(count = "ifelse(cyl == 4, 1, -1)", "hp <= 120"))
side_plot(mtcars, "cyl", c("hp <= 120", "~ - wt / cyl"))
side_plot(mtcars, "cyl", c("1", "-1"))
}
