% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_fig.R
\name{square_fig}
\alias{square_fig}
\title{Change the aspect ratio of a plot}
\usage{
square_fig(p, aspect_ratio = 1)
}
\arguments{
\item{p}{: a ggplot object.}

\item{aspect_ratio:}{aspect ratio of the panel, Default = 1, which will make
x and y the same length and hence the figure a square.
A value > 1 will make y-axis longer than x-axis.
A value < 1 will make x-axis longer than y-axis.}
}
\value{
a ggplot object of square shape by default or a rectangular
        shape when user decides not to use 1 as aspect ratio.
}
\description{
\code{square_fig} takes a ggplot object as input and changes its shape to
a square by default. To change the plot to different degrees of rectangular
shapes, user can specify an aspect ratio different from 1.
}
\examples{
inst/examples/ex-square_fig.R
}
