% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Test if some observed data are normally distributed.}
\usage{
test_normality(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(varname, linew = 0.7, xlab_left = varname, title_left,
               title_right, subtitle_left, subtitle_right,
               caption_left, caption_right, xscale_left, ...)}
\itemize{
     \item varname. String, name of a continuous variable. Its empirical CDF
     will be plotted along side its normal probability plot.
     \item linew. Number, width of the line. Default = 0.7.
     \item xlab_left. String, x label of the left figure. Default is varname.
     \item title_left. String, title of the left figure.
     \item title_right. String, title of the right figure.
     \item subtitle_left. String, subtitle of the left figure.
     \item subtitle_right. String, subtitle of the right figure.
     \item caption_left. String, caption of the left figure.
     \item caption_right. String, caption of the right figure.
     \item xscale_left. String, scale of x-axis of the left figure. Possible
     values are described at \code{\link{scale_axis}}. There's no parameter
     to change the x-scale of the right figure because the x-axis of the
     standard normal probability plot always goes from -3 to +3 with 1
     increment at a time.
}
}
\description{
The normal distribution is the most commonly used probability distribution.
Many real world data such as height can be modeled via a normal distribution.
\code{test_normality} takes in a data frame and returns a function for making
ggplot2 type of CDF and the standard normal probability plots side by side
on one canvas of any continuous variable from the data frame. If normal
probability plot is linear, the observed variable is normally distributed
with mean and standard deviation equal to the sample mean and sample standard
deviation.
}
\examples{
inst/examples/ex-test_normality.R
}
