% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Create a function for drawing publishable ggplot2 CDF plot and normal
probability plot (comparing the sample quantiles to the quantiles of the
standard normal distribution) side by side on one canvas. The resulting
figure is often used to test normality of the sample distribution.}
\usage{
test_normality(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(varname, xlab_left = varname, title_left, title_right,
               caption_left, caption_right, font_size = 14)}
\itemize{
     \item varname. String, name of a continuous variable. Its empirical CDF
     will be plotted along side its normal probability plot.
     \item xlab_left. String, x label of the left figure. Default is varname.
     \item title_left. String, title of the left figure.
     \item title_right. String, title of the right figure.
     \item caption_left. String, caption of the left figure.
     \item caption_right. String, caption of the right figure.
     \item font_size. Overall font size. Default = 14. The font size of the
     axes and legend text is a fraction of this value. It controls both plots.
}
}
\description{
\code{test_normality} takes a data frame as input and returns a function for
drawing CDF plot and normal probability plot of any continuous variable
from the data frame.
}
\examples{
inst/examples/ex-test_normality.R
}
