% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Create a function for drawing publishable ggplot2 histogram and
Q-Q normal plots side by side on one canvas. The resulting figure is often
used to test normality of the sample distribution.}
\usage{
test_normality(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
\code{function(varname, binw = NULL, bins = 30, detrend = TRUE,
               font_size = 14, title_hist = NULL, title_qqplot = NULL)}
\itemize{
     \item varname : string, name of a continuous variable. Its
                     empirical distribution will be plotted as histogram,
                     and its normality will be tested and a qqplot will
                     also be plotted.
     \item binw    : width of the bins. Can be specified as a number, or
                     a function that calculates width from x. Default = NULL.
     \item bins    : number of bins. Overridden by binw. Default = 30. You
                     almost always want to change the defult. Andrew Gelman
                     recommends to choose a sufficiently large number to
                     show how raw data are distributed without any smoothing
                     as happens when estimating density. See this blog post
                     for details: http://andrewgelman.com/2009/10/23/variations_on_t/
     \item detrend : logical, Should the resulting Q-Q plot be detrended?
                     If TRUE, it'll be detrended according to the reference
                     Q-Q line. This procedure was described by Thode (2002),
                     and it reduces visual bias caused by orthogonal
                     distances from Q-Q points to the reference line.
                     Default = TRUE.
     \item font_size : overall font size. Default = 14. The font size of the
                       axes and legend text is a fraction of this value.
     \item title_hist   :  string, title of histogram. Default = NULL.
     \item title_qqplot :  string, title of qq normal plot. Default = NULL.
}
}
\description{
\code{test_normality} takes a data frame as input and returns a function for
drawing histogram and Q-Q normal plot side by side of any continuous variable
from the data frame.
}
\examples{
library(ezplot)

f = test_normality(iris)
f("Sepal.Length")
f("Sepal.Length", bins = 50, detrend = F)

library(dplyr)
f = test_normality(films \%>\% filter(made_money == "yes"))
f("boxoffice", font_size = 10)

# take log of boxoffice first and then re-run
df = films \%>\% filter(made_money == "yes") \%>\% mutate(logbo = log(boxoffice))
f = test_normality(df)
f("logbo")
f("logbo", bins = 100, detrend = F)
}
