% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{tile_plot}
\usage{
tile_plot(data, x, y, z, facet_x = NULL, facet_y = NULL, size = 20,
  ncol = NULL, labels_x = NULL, labels_y = NULL, labels_z = ez_labels,
  palette = ez_jet, reorder = c("facet_x", "facet_y", "x", "y"))
}
\description{
tile_plot
}
\examples{

tile_plot(mtcars, "factor(cyl)", "factor(am)", "mpg")

library(ukbabynames)
library(dplyr)
ukbabynames \%>\%
  mutate(name = toupper(name)) \%>\%
  filter(substring(name, 1, 1) \%in\% LETTERS) \%>\%
  tile_plot("substring(name, 1, 1)", "substring(name, 2, 2)", "n")
}
