% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{tile_plot}
\usage{
tile_plot(data, x, y, z, facet_x = NULL, facet_y = NULL, size = 12,
  ncol = NULL, labels_x = NULL, labels_y = NULL, labels_z = ez_labels,
  palette = ez_jet, reorder = c("facet_x", "facet_y", "x", "y"))
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character. Evaluates to a column.}

\item{y}{A named character. Evaluates to a column.}

\item{facet_x}{character. Evaluates to a column.}

\item{facet_y}{character. Evaluates to a column.}

\item{size}{theme size for \code{use_theme()}. Default is 20.}

\item{palette}{Colour function.}

\item{reorder}{A character vector specifying the group variables to reorder.
Default is \code{c("group", "facet_x", "facet_y")}.}
}
\description{
Creates tile plots.
}
\examples{
tile_plot(mtcars, "factor(cyl)", "factor(am)", "mpg")
tile_plot(ez_data(), "year", "char", "value", "fct", "num", reorder = NULL)
}
