% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{tile_plot}
\usage{
tile_plot(
  data,
  x,
  y,
  z = c(Count = "1"),
  facet_x = NULL,
  facet_y = NULL,
  size = 11,
  facet_ncol = NULL,
  labels_x = NULL,
  labels_y = NULL,
  labels_z = ez_labels,
  zlim = function(x) c(pmin(0, x[1]), pmax(0, x[2])),
  palette = ez_jet,
  reorder = c("facet_x", "facet_y")
)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{A named character value. Evaluates to a column.}

\item{y}{A named character value. Evaluates to a column.}

\item{z}{A named character. Evaluates to a column and is mapped to the fill
colour of the tiles.}

\item{facet_x}{A character value. Evaluates to a column.}

\item{facet_y}{A character. Evaluates to a column.}

\item{size}{theme size for \code{use_theme()}. Default is 14.}

\item{facet_ncol}{Option passed to ncol argument in \code{facet_wrap} or
\code{facet_grid}. Default is \code{NULL}.}

\item{labels_x}{label formatting function}

\item{labels_y}{label formatting function}

\item{labels_z}{label formatting function}

\item{zlim}{argument for \code{scale_fill_grandientn(limits = zlim)}}

\item{palette}{Colour function.}

\item{reorder}{A character vector specifying the group variables to reorder.
Default is \code{c("group", "facet_x", "facet_y")}.}
}
\description{
Creates tile plots.
}
\examples{
\dontrun{
library(tsibbledata)
library(dplyr)
nyc_bikes \%>\%
  mutate(duration = as.numeric(stop_time - start_time)) \%>\%
  filter(between(duration, 0, 16)) \%>\%
  tile_plot(c("Hour of Day" = "lubridate::hour(start_time) + 0.5"),
            c("Ride Duration (min)" = "duration - duration \%\% 2 + 1"))
}
}
