% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/web_display.R
\name{web_display}
\alias{web_display}
\title{Enlarges a ggplot2 plot for web display}
\usage{
web_display(p, axis_text_size = 20, axis_title_size = ggplot2::rel(2),
  axis_text_angle_x = 0, axis_title_vjust_y = 1, legend_pos = "right",
  mar_top = 1, mar_right = 1, mar_left = 1, mar_bottom = 1)
}
\arguments{
\item{p}{ggplot object}
}
\value{
a ggplot object
}
\description{
\code{web_display()} takes a ggplot object as input, and enlarges its axis
tick text, axis label text, legend title and text, and distance between the
axis labels and the axis to make it look good for web display.
}
\examples{
# make some data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
p = barplt("student", "grade", "student", xlab="Student", ylab="Grade")
print(p)

# make the barplot web-ready
web_display(p)
web_display(p, legend_pos="top")
web_display(p, legend_pos="none")
web_display(p, legend_pos="bottom")
}

