% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_display.R
\name{web_display}
\alias{web_display}
\title{Enlarges a ggplot2 plot for web display.}
\usage{
web_display(p, axis_text_size = 16, axis_title_size = ggplot2::rel(2),
  axis_title_vjust_y = 1, legend_pos = "right", legend_title = T,
  mar_top = 1, mar_right = 1, mar_left = 1, mar_bottom = 1)
}
\arguments{
\item{p}{A ggplot object.}

\item{axis_text_size}{A number, size of the tick text. Default = 16.}

\item{axis_title_size}{A number, size of the axis labels. Default = rel(2).}

\item{axis_title_vjust_y}{A number, distance between y label and y-axis. Default = 1.}

\item{legend_pos}{A string, location of the legend. Default = "right".}

\item{legend_title}{Logical, whether to display legend title. Default = TRUE.}

\item{mar_top}{A number, top margin of the whole plot. Default = 1.}

\item{mar_right}{A number, right margin of the whole plot. Default = 1.}

\item{mar_left}{A number, left margin of the whole plot. Default = 1.}

\item{mar_bottom}{A number, bottom margin of the whole plot. Default = 1.}
}
\value{
A ggplot object
}
\description{
\code{web_display} takes a ggplot object as input, and enlarges its axis 
tick text, axis label text, legend title and text, and distance between the 
axis labels and the axis to make it look good for web display.
}
\examples{
# make some data
df = read.table(header=TRUE, text='
student grade
Joe 90
Mary 75
Alex 50')

# draw a barplot
barplt = mk_barplot(df)
p = barplt("student", "grade", "student", xlab="Student", ylab="Grade") 
print(p)

# make the barplot web-ready
web_display(p)
web_display(p, legend_title=FALSE)
web_display(p, legend_pos="top")
web_display(p, legend_pos="none")
web_display(p, legend_pos="bottom")
}
